[roll] Roll fuchsia [kernel][lockup][ddk] Downgrade SMC overrun from oops to warning
We're seeing a fair number of these in the 10-100ms range.  Prior to
this change an overrun of 10+ ms would trigger a kernel OOPS
which (via tefmocheck) will cause tests running at the time to be
classified as failures.

Downgrade the OOPS to a warning and remove the message-specific
tefmocheck so that we don't unnecessarily fail tests and waste infra
resources.

Also, decrease the rate limiter period from 10s to 1s so that we'll
get better visibility into repeated overruns that exceed the
threshold.

Original-Bug: 71167
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495899
Original-Revision: bd9e83c9f2ad94fbac0eef2f72c97f39aa5bb120
GitOrigin-RevId: a6b6d2292dd441c1b3d4481bb7667c1701f2c5ba
Change-Id: I2ce1a4d534eca9bd889b1d1534734789486d9db2