[roll] Roll fuchsia [tracer] Fix for nonexisting depfile
Action tracer expects the depfile to exists, which
may not be true in many cases, so it starts complaining
of invalid writes when creating depfiles. So fix the
action tracer such that, even if it doesn't exist,
allow writing the depfile.

Original-Bug: 71561
Test: fx build/tracer

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496101
Original-Revision: 655f21d28270079f2404da438bbb9c86530f60a3
GitOrigin-RevId: 4b9cff109fea90d3aa971bdc320468aefd5a0f99
Change-Id: I25843600270c33bdae6247cf134cbf5a7c7aaf1b