[roll] Roll fuchsia [scenic] Fixed image creation bug.
Since image creation commands can be enqueued on a Scenic session
without immediately presenting, this can create a large enough
window for the buffer collection that the image is created on to be
deregistered in the meantime.

The buffer collection, however, doesn't immediately get freed, which
means that it can still be used for the late image creation. This
patch adds a testcase for this issue and adds a fix for it.

fx test scenic_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495960
Original-Revision: c9d0ea4366df34c57cc00d2a4d7b93b053ce6cde
GitOrigin-RevId: b6ed0fb099cbaac91bf10fe91af0d9fbcb129f20
Change-Id: I025fc7e692286dd71f09e530aa9b81ce49c142c4