[roll] Roll fuchsia [mdns] provide flow control for fuchsia.mdns.PublicationResponder
When a client of mdns service wants to publish a service, the client
implements fuchsia.mdns.PublicationResponder and sets up a channel so
the service can call that implementation. In network environments
with many nodes searching for the service type in question and under
heavy CPU load, the channel carrying PublicationResponder messages
can overflow as one the parties fails to read promptly enough from
the channel. This results in panics at either end of the channel, at
which point the device ceases to function as a cast device.

To address this, this CL adds flow control for the single method of
PublicationResponder. The class responsible for this relationship on
the mdns service side now queues requests internally if there would
otherwise be more than two such method calls in progress. This
prevents the channel from overlfowing, either with requests or
replies, under any conditions.

The throttling of these requests may slow the mdns's network
responses by a few milliseconds, but this will not affect the user
experience noticeably. The maximum number of oustanding calls is
limited to two (rather than one) so the operation of the client and
service are not completely serialized, allowing for parallel
execution.

Test: new unit test for PublicationResponder adapter
Original-Bug: b/177137740

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495205
Original-Revision: b0a551e49c9dbbac34897378fe7c48e57befd1ce
GitOrigin-RevId: b35ba101409b2bf0585cd28fd48e114efdfe8b16
Change-Id: I90794c5be218722e769cdbba57dfa7339c63a029