[roll] Roll fuchsia [fuzzing][fidl] Dep //src/lib/fuzzing/fidl:tests
With changes landing upstream in compiler-rt for remote fuzzing, the
//src/lib/fuzzing/fidl library needs to be reworkd. But first, the
library has bit-rotted due to a missing dep on its tests. Rather than
fix code to be replace, this CL adds a dep but makes the group target
empty. As code is replaced, it will be added to target.

Original-Bug: 71012
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491326
Original-Revision: 6e145725feec04ba62d4cc445e18ffac975f96cd
GitOrigin-RevId: e0a1460cd37141128b476754148d27e01d825aaf
Change-Id: I90236a92e1e483787cf3de798c0dfd91614c53cf