[roll] Roll fuchsia [gigaboot] add LZ4 decompression capabilities
Adds functionality to decompress LZ4 images. This will be necessary for
FVM flashing, since the images are large enough that decompression saves
substantial transfer time.

LZ4 decompression must be done in chunks in order to allow time to
service other interfaces - in particular, the network. This
decompression will take long enough that if we did it all in one go, we
would fail to respond to ICMP and/or fastboot packets for too long and
the connection would drop.

This CL adds the device-side support and host-side tests, but nothing is
actually using it yet. Follow-up CLs will hook it into the fastboot
flashing routines.

Original-Bug: 68692
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495995
Original-Revision: daa2dd1258b42c94dd0f1917299e6eeabfc2a1ce
GitOrigin-RevId: b4a897116135d0926d7aee38dcb60aec2bb03186
Change-Id: Id3067556a46228134d513f5c35d59ad12f9efc79