[roll] Roll third_party/go [fidl][go] Part 1/2: Move conformance test file listing to Go repo
This is meant to be combined with:
https://fuchsia-review.googlesource.com/c/fuchsia/+/496140

Currently, it isn't clear that you need to update a file list to get the
go conformance tests to pass. Additionally, you need to do a multistep
migration to add a file because you need to first add the file to the go
repo, then add the filename to fuchsia.git, then use the file in the go
repo.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/496620
Original-Revision: c59f469233adb3985e252fe3223912f1413129e5
GitOrigin-RevId: c95574f0f68b1322750fe3831f024a1e852bf16a
Change-Id: Id5c18b1401f64c0f6aca5953e9aa7d10a6ae02fb