[roll] Roll fuchsia [bind] Decode new bytecode into a bind program
Introduce a DecodedProgram struct that reads in
bytecode and create a symbol table and list of
instructions from it.

When the DecodedProgram receives the bytecode,
it verifies the headers and section. After that,
it splits the sections into separate vectors and
decode them seperatly.

This CL also moves common enums and constants
into a separate file.

Original-Bug: 69608
Test: Added unit tests in decoded_bind_program.rs
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/494819
Original-Revision: ed55073d27753d9ba5eeed2026839a6d63e6d0cc
GitOrigin-RevId: 5a4e74c5cf0ad12f32790a76c04c6b9263d05810
Change-Id: Ib3df8d46ddb950ddd4f27ca99a21bc5a0ce51d72