[roll] Roll fuchsia [kernel][lockup] Reboot when fatal thresholds are exceeded.
Create a crashlog and reboot the system with a reboot reason of "SW
WATCHDOG" when any of the lockup conditions' configured fatal
thresholds are exceeded.

Also, alter the heartbeat check-in behavior to pet the watchdog iff
the heartbeat condition has a configured fatal threshold.  With this
change, we should never see a HW WDT firing as long as there are
active cores who are still checking in on each other.

With these changes, we should be in a position to start collecting
data from production field devices which are hitting the HW WDT.  Now,
instead of hitting the HW WDT, they should end up tripping the SW
WDT (lockup detector) level instead, generating a crashlog containing
details for us to analyze.

Original-Bug: 70724
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496819
Original-Revision: d42abf097c24bcfcb0f6e5d24382ac1a1e2e074d
GitOrigin-RevId: f65448a2b0aeac4c31e1839f13926be13aaab57b
Change-Id: Ia77f6368de6736e3e3bf8b873ab29506a00b6405