[roll] Roll fuchsia [fdio] Remove dupcount
This changes the behavior of `fdio_unbind_from_fd` to always consume
the passed file descriptor. The new behavior matches the behavior of
`fdio_get_service_handle`. `fdio_fd_transfer` is likewise modified.

`fdio_get_dupcount` was being called in 3 places:

- `fdio_unbind_from_fd`: this function is also checking
`fdio_is_last_reference`, which is a superset of checking the dupcount,
so checking the dupcount is unnecessary.
- `__libc_extensions_fini`: this function was guarding `fdio_release`,
which internally checks `refcount`, so checking the dupcount is
unnecessary.
- `fdio_get_service_handle`: this function doesn't check `refcount`, but
calls `fdio_ops_t::unwrap` whose documentation indicates concurrent
operations must not be ongoing when it is called. This function is
rewritten in terms of `fdio_unbind_from_fd` since it is virtually
identical.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496582
Original-Revision: 637724388d3cb33485a2127187069555b7138399
GitOrigin-RevId: 517a381f2689c9da98224127d99dd2079a670eda
Change-Id: Ic3787069c9d70b418de616ce855b009c8e03e6c2