[roll] Roll fuchsia [blobfs] Implement VmoRead for new pager
Plumbs the VmoRead request for the new pager from Blob through to the
UserPager. Minor changes are made to UserPager and PageWatcher (the old
paging infrastructure) to allow this to happen in the same way whether
the new pager is enabled or not.

Make VMO parameters const. The zx library sues const on a VMO to mean
the C++ object, not the udnerlying system object, is constant, while the
old pager library assume const meant both were constant. This mismatch
caused some problems when hooking up to the blobfs code.

The new pager code is not currently compiled by default and does not yet
work. The next step will be creating the paged VMO properly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497719
Original-Revision: cda3b63613c6680198e0d4ca24d535674f57a586
GitOrigin-RevId: 718aedfcd46a0f6892c3eb0a419150227b55a6fb
Change-Id: I3624ea42952d0c2936ba831929919e4d9b93ad55