[roll] Roll fuchsia [fidl][go] Part 2/2: Move conformance test file listing to Go repo
This is meant to be combined with:
Idb87484f2ea8e672ffd2fc940225aedaa834f546

Currently, it isn't clear that you need to update a file list to get the
go conformance tests to pass. Additionally, you need to do a multistep
migration to add a file because you need to first add the file to the go
repo, then add the filename to fuchsia.git, then use the file in the go
repo.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496140
Original-Revision: 2970a3983041dcb421bd65992c8337f90848e038
GitOrigin-RevId: 9a3c1e64778baf72831769bcf6bc3f3eb693e478
Change-Id: I443a95e80179ee3ba8a85c848294ed05deadbd5c