[roll] Roll fuchsia [fidl][test] Add missing dart dependency
Several tests under dart_bindings_test depend on
`test/test/server.dart`, while it's not explicitly listed as a source
to `bindings-test-component`. This mean if `test/test/server.dart` is
changed, it won't correctly trigger a rebuild of this test component
when building incrementally.

To fix it, I wrapped `server.dart` in a library and set it as a
dependency of the test component.

NOTE: At first glance a simpler fix would be to add `server.dart` to
`sources` of the test component directly. However that didn't work
because `dart_test_component` expects a `main` function in all of its
sources and uses that to generate a `main.dart` for test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497485
Original-Revision: 3a846caddbf8e3c45e38cfb74b1357271ebe5c59
GitOrigin-RevId: cac89532eb573b97bd23f87bc3f118fa2a79718c
Change-Id: I05a49343ac807ef2ca6cc24c2b96be69c1a5fa84