[roll] Roll fuchsia [media][audio_core] Prevent SetReferenceClock race
In both AudioCapturer and AudioRenderer, a boolean ensures the reference
clock is only set once. The boolean is not protected by a mutex, so a
race can occur when concurrent calls to SetReferenceClock are made. This
CL protects |reference_clock_is_set_| with a mutex.

Test: MULTIPLY: audio_core_unittests
Test: MULTIPLY: audio-core-api-tests
Test: MULTIPLY: audio-core-api-pipeline-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483340
Original-Revision: bcfbfe80b115308cf1f906b43ed7290d19c0897f
GitOrigin-RevId: be527eef8a86025072e93cf0dd55c048afd2e30c
Change-Id: I36f029d2019c26107d76d80bd4c7278c19c4e1d2