[roll] Roll fuchsia [blobfs] Move code out of blob files.
Moves Blob::WriteInfo out of the header and into the .cc file which
makes the Blob header a bit easier to follow. This struct only needs to
be forward-declared and is used only in blob.cc.

Moves the "Producer" classes out of blob.cc and into their own file.
This is a non-trivial amount of code and allows the compliex blob.cc
file to be a bit simpler.

There should be no behavior change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496764
Original-Revision: b669595b31a3adeb2c0142c8823c4f97464e6a4f
GitOrigin-RevId: 149c4d0dda1d8bbe6d4003356aa7e6005dd68a34
Change-Id: I8e41009389978a5f207664528302a6938966e09d