[roll] Roll fuchsia [virtio] Migrate to zx_system_get_page_size
Migrates usages of page size constants to zx_system_get_page_size

The virtio rings are not required to be sized at precisely the system
page size, the only requirement is that they are contiguous in the
physical view of memory. As such they could be sized smaller. However,
since the existing implementation chose to tie the ring size to the
system page size and not make it smaller, this preserves that choice.

Original-Bug: 67899

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497998
Original-Revision: 69dc14ab317e768be438eb001a2b74c5c1516d3f
GitOrigin-RevId: 7174fdfd1a1c97d768f3193c9906bd51c5919ae7
Change-Id: I915b460e46fa64226503b3228f06fdda29172344