[roll] Roll fuchsia [subprocess] Detach subprocess's stdin from the terminal
I don't fully understand why, but unless I do this, subprocesses are not
able to receive stdin piped from the parent processes, even when
`cmd.Stdin` is set to `os.Stdin`.

Also, set `os.Stdin` as the default stdin for commands that don't
specify a stdin so that tools need not explicitly pass `os.Stdin` when
invoking subprocesses that read from the parent's stdin.

This is necessary for the interactive `goma_auth.py login` to work when
called by the new version of `fx set`.

Original-Bug: 68465
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498038
Original-Revision: 4404860161eaaac4f18030f91ebcd683ea969839
GitOrigin-RevId: d2bb276f2f930e7ff3ed50f46899888b684a039c
Change-Id: Id00237dce9626f430d6d5fd2b8797db9bf1b1c88