[roll] Roll fuchsia [fx] Rename fint_set feature to legacy_set
And invert its meaning. Whereas `fint_set` defaulted to false and made
`fx set` use the new implementation if true, `legacy_set` defaults to
true and `fx set` will use the new implementation if it's false.

The intent of the renaming is to make it more obvious and memorable, as
"fint" is not a widely known term. It should be safe because the new
feature has not yet been announced.

Original-Bug: 68465
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497980
Original-Revision: 97dbf8d07df085e06f8e66315d22ba708fa873db
GitOrigin-RevId: fde1b871cdbbb1bf39faa0b90b8ed49b13ef8d96
Change-Id: Ibce592fa1c869e86fc226a0fc849314b93a0efba