[roll] Roll fuchsia [codec] Keep Magic used diagnosing fxbug.dev/67247
This Magic class is useful enough for debugging problems
that it seems worth keeping.  At the moment we don't
have any classes with a Magic<> member in them.  The
intent is to use a Magic<> member when we're trying to
narrow down anything that looks like corruption or
lack of type safety.

BUG: fxbug.dev/67247
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497000
Original-Revision: 208f22570036f681d038d05225cceec12053ec82
GitOrigin-RevId: 0e4c375a22c3fd4247deccdebb9c868ef0551a86
Change-Id: I14e939baaf2a5c8c38a70139ba790ba7ae1b1226