[roll] Roll third_party/go [go][comformance] Fix handle replace + custom deep equals
Due to a bug in go decode (see encoding_new.go), handles
were being replaced with handles with reduced rights but
the replaced handle was being used rather than the new
handle produced from handle.Replace.

Once the fix was made, the decode tests that perform
handle replacements started failing because
reflect.DeepEquals was being used to replace the result.
This CL fixes the bug and creates a new deep equals that
does not perform handle equality checks. This is not
ideal, but a multi-step migration is needed to generate
information that can be compared in the deep equals
function so it is useful to get the fix in first.

Original-Bug: 71765
Original-Bug: 41920
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/496500
Original-Revision: 172637d6d28c3bf3eb40bc13b4c3408303f972bd
GitOrigin-RevId: 1f21ef84c5a14ed4d5b6d346e4450a04c1f05096
Change-Id: I73d89bb1baae8ae4c1f93f4ec080f64a3ed23c23