[roll] Roll infra/recipes [cherry_pick] Add disambiguation for duplicate project names
There is an edge-case in integration.git where there exists a duplicate
project name for two different remotes, in two different manifests. We
cannot change this edge-case to conform to project name uniqueness, so
the cherry-pick recipe needs a mechanism to disambiguate.

The project's remote should work for this case, and is something that
can be derived and passed by release-tool.

Original-Bug: 71788
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/497881
Original-Revision: 5ef902c03a373e49341eb8da8ec272446f92fd38
GitOrigin-RevId: 766734df0ab7f6471d60725eba5a40783f6cce0d
Change-Id: Iadbd02b4e7592846464ac7544cb851cae0b96a4e