[roll] Roll fuchsia [libtriage] Support Not(Missing(...)) for unknown types.
Some values (like histograms) cannot currently be handled by libtriage,
however, we want to be able to assert that they are at least present.

This change adds a new MetricValue::NotMissing alternative that may be
used for types that are present but not understood.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490865
Original-Revision: e8d500603f1f655dae52cdbd9076183574ef3130
GitOrigin-RevId: 6c0e502662cbb0d077f47a98dbfd35c5912d8149
Change-Id: I5ae6bdd354466eeecbaa07d8b75548b0d69772aa