[roll] Roll fuchsia [wlan] Immediately identify an AP as Open if the privacy bit is not set
This change removes the possibility that an unusual network
configuration could be identified as WPA2 when the privacy bit is not
set. This privacy bit itself conventionally determines if a network is
Open or not.

The mocked Wpa2NoPrivacy security type is removed and no longer
supported. This security type is a corner case in IEEE 802.11-2016
that we do not test support for, and it should not be supported since
it implies a WPA2 connection that is not encrypted.

Test: Existing test coverage is sufficient.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470421
Original-Revision: 2c30105be8e01185392f1002afc1ef5c40b14ce9
GitOrigin-RevId: 0282e7bb65cdeacd05c7480c76c48e0abd889803
Change-Id: I3a6129b7de2bd9dd7e3e6546bfc5e6dd0cdb72d8