[roll] Roll fuchsia [docs][components] Cover logging to debug log
Tests or workflows which depend on logs coming from the debug log
require explicit work in v2. Watching the debug log is known to
be a pattern for some integration tests which run something on
an emulator and then watch for logs.

Original-Fixed: 70512
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/492962
Original-Revision: 68dc3063480dbac9d8090e693d05f6493d28a4ae
GitOrigin-RevId: 7850b83d4b18867f65c70957b3e1ee5e912ea523
Change-Id: If977063f79047e7a7e3f72ead1c20f78f0579176