[roll] Roll fuchsia [fidlc] Conditionally remove support for protocols as types
This CL modifies fidlc so that under the new syntax, protocols
cannot be interpreted as types. To do this, it refactors
CompileTypeConstructor to validate the "sort" (see
ExpectedSort in flat_ast.h) of the thing it is compiling.

In the old syntax, CompileTypeConstructor allows either types
or protocols, whereas in the new syntax it only allows types.
In either case, we never allow using services as a type. Also,
`CompileService` leverages this to validate that its members
are protocols.

Test: fidldev test fidlc
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486197
Original-Revision: 450d5be45083c16c32d43991a66376d0463da187
GitOrigin-RevId: 92000ba15a3d1df05746018d46616bb0dfee73e3
Change-Id: Ic1c8edbca6d10228f46ee3ca9c781732a0785174