[roll] Roll fuchsia [build][tracer] Unwrap known wrapper commands
The action-tracer ignore logic uses the name of the underlying script,
but we are not interested in command-wrappers that may appear before it.
This peels apart wrapper commands to get to the tool of interest.

This is needed to make the ignore logic works in the presence of command
wrappers, e.g. output_cacher.py.

Test: (enable tracing in fx set; fx build)
Test: fx test build/tracer

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497610
Original-Revision: 9360165b2b43baccac2e2ad797e165901af2f776
GitOrigin-RevId: a0b1712c7a2e6d390e60d61fdfffb4f16602e44a
Change-Id: I33497516469d6f207c58cd9b2bb00192237bb9f2