[roll] Roll fuchsia [runner] Don't set a pgid when forwarding os.Stdin to subprocess
I don't fully understand why, but unless I do this, subprocesses are not
able to receive stdin piped from the parent processes, even when
`cmd.Stdin` is set to `os.Stdin`.

This is a problem for interactive goma login when triggered by a
subprocess of `fx-set`.

This is effectively a combined revert+reland of fxrev.dev/498038, which
attempted to accomplish the same thing using a different approach.
Conditionally setting `Noctty` solved the problem on Linux but not on
Mac.

I've tested this approach on both Linux and Mac and it seems to work.

Original-Bug: 68465
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498141
Original-Revision: a0f16584ce9db68a4b667f82d4da2473c360fca9
GitOrigin-RevId: cb8600402ac628c2d1d52f4bcc10a4b9958faf88
Change-Id: Ia86dfa7340d5a58d8cf510727669fd8c938671b3