[roll] Roll fuchsia [kernel][crashlog] Fix uptime units in the crashlog startup banner.
Fix the uptime units printed out in the crashlog startup banner.
Before this, the banner reported uptime in nanoseconds, but the
message said that the units were "ms", which is incorrect.  Now, we
report the values in seconds, but with 3 digits of msec resolution as
well.  This is the format used for presenting timestamps in kernel log
messages, so hopefully it will be easy for folks to read and reason
about.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497649
Original-Revision: 64986badfff3a0fe648c60c408fe3f98643ad3d6
GitOrigin-RevId: 3f33f3c908aa03f2ccc1c8f688cae955756b545a
Change-Id: I06192ef2a87f090abaf6e4629f3991bdb8c157c3