[roll] Roll fuchsia [debugger] Show thread context on global pause.
Outputs the current thread context on a global pause. This wasn't
getting output in the common case of pausing immediately after attaching
because there was no current thread (the threads had not be synced from
the debug_agent before the pause was issued).

This change always outputs the current thread after the command
completes which avoids the issue, and also is more consistent if the
current thread changes during the time the pause was being issued.

Allow "global" annotations for "pause" and "continue" commands. This is
technically redundant but I found myself wanting to be explicit and was
surprised when this issued an error.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499060
Original-Revision: d1d0d6a9a4ca144a99a90c934757075c3003982f
GitOrigin-RevId: 1ba32f548f8f9c2cc1f8a8b25a68feb1ca5ccdb4
Change-Id: I51dbd3242ab8b181272482a2d0d1875594311362