[roll] Roll infra/recipes [update-petal] Add override for target manifest
Add a feature to optionally override the target manifest to update,
bypassing the logic to use a Jiri project query to infer the target
manifest based on the target project.

This is useful for updating projects which are not part of the default
checkout, and thus invisible to Jiri.

Test: https://ci.chromium.org/swarming/task/52367e593869ee10?server=chrome-swarming.appspot.com
Original-Bug: 71916
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/499280
Original-Revision: a1bed6f84097d83f6eec62262a437b2800c291fc
GitOrigin-RevId: 6faa622c4ae9e944af5dec0677cf2fea8c0f5cf4
Change-Id: I9f638bca69756e093ac3e001fb8ea89da6913a15