[roll] Roll fuchsia [fint] Create a stub `fint build` command
For now it's a no-op that just prints out the static and context specs
the caller passes to it. I'm working on integration with recipes, and
this will at least make it possible to call `fint build` from recipes
even if it doesn't produce any of the expected build artifacts.

I'll actually implement the build logic on follow-up changes.

Rather than duplicating a bunch of code between the set and build
commands, I extracted a `BaseCommand` type that holds all the common
logic.

Original-Bug: 67861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499478
Original-Revision: 232e454d4d42638ed10574abc5d55855a2132531
GitOrigin-RevId: 7a552d9b7a1f6d329063ca1430952037a3d1e4df
Change-Id: Ief729f1c865d19f1cbb6e4d8ccba15443023e8da