[roll] Roll infra/recipes [build] Implement support for `fint build`
If the `fint_build` arg is True, invoke `fint build` instead of `ninja`.
This won't work yet because `fint build` is only a stub so far
(fxrev.dev/499478), but it will give me a platform to iterate on more
quickly.

It's slightly hacky, but I decided to pass the necessary artifacts
(namely the fint executable and the two specs) into `fint build` by
attaching them to the existing `GNResults` struct. This is a bit of a
hack because these files are not truly part of the "GN results", as
they're created before GN even runs. However, it keeps the code fairly
clean with minimal plumbing, and ensures that `fint build` will always
use the exact same fint executable and specs as `fint set`, and won't
unnecessarily rebuild fint or re-write the context spec.

Original-Bug: 499478
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/498584
Original-Revision: b3bd0e7bd3897e57549dab7a5356a0e96242c2ad
GitOrigin-RevId: a22060039ec58e3f8ffd87db2ff544b632676331
Change-Id: Iee44dac40b8c44170a8d7ba7d7f64be63eecd9e9