[roll] Roll fuchsia [kernel] Fix mexec case when RAMDISK needs to be relocated
The mexec code moves the RAMDISK image to a different address if
forced by a kernel command line option or if required to avoid
overlap with the kernel image.  In the latter case, it was
setting up to do the copy and then acting like it had done the
copy but failing to actually do the copy.

Original-Bug: 32414
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499600
Original-Revision: 3d46cfadffe9754a567d45b7d2d80ed03564fa35
GitOrigin-RevId: 4957c4f27ebb416f21bfbbfbcf3dacc27acafaa8
Change-Id: Ifbb6a4c0f300cb4d8ce59d40d69c741f9d5e8df1