[roll] Roll fuchsia [kernel][wdt] Fix a hypothetical issue with a suppressed WDT
Even if WDT petting is suppressed, make sure to schedule the next pet
timer relative to just before when we _would_ have pet the dog.  We
want the last_pet_time to accurately reflect when we last actually pet
the dog, but if we schedule our next timer based on last_pet_time, and
petting is suppressed, this could cause us to end up continuously
scheduling a timer in the past, effectively lockup up a core.

Right now, we only ever suppress dog petting during the generation of
a crashlog during a SW WATCHDOG event.  This is our failsafe to ensure
that the system _will_ reboot, even if the crashlog generation hangs.
Even so, it would be best to not end up simply spinning another core
in timer IRQ processing while this happens.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498459
Original-Revision: 1be50a559e6443859445e396162ea11b4630baed
GitOrigin-RevId: e0c35f27c823686fc367a1b52e4d6378af01bb3e
Change-Id: I98f3fa9b3a4fe434669adfe3f062c87bfd36875c