[roll] Roll fuchsia [settings] Expose service messagehub to env creator
In order to watch the service MessageHub via a broker,
the creator of the environment needs to have access to the message
hub. This CL changes the return so that it returns the message hub
along with the nested environment.

Original-Fixed: 71062
Test: fx test -o setui_service_tests
Multiply: setui_service_tests

Test added to check that the message hub returned from environment
creation is the same instance as the one in the rest of the service.
Uses a test agent to receive a broadcasted message over the message hub
returned by environment creation and verifies that the test agent can
receive it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/492019
Original-Revision: 87e31c27670ca5b29da9ae28f44edd6617ccaf7d
GitOrigin-RevId: 66c53f0a2556632784f292663e1fe9e27d43d25d
Change-Id: If9293d8bb491b110359e91880fcb9c5306e557ba