[roll] Roll fuchsia [timekeeper] Increase some warnings to errors.
Previously we'd tended to use warnings for conditions that really
shouldn't be occuring but were not our fault and which we could recover
from.

However, RFC-0003 states "Note that the bug might not be in the code
issuing the log statement. The bug may be in one of the callers or one
of its dependencies." Given that and given that errors now have quite a
bit more tooling to help track and isolate them I'm bumping a few
conditions up from warning to error.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493010
Original-Revision: e98cc10df6ed75a3968659c2b9dc20580dfca94d
GitOrigin-RevId: 3df05d7eb518e08d0c89c3880029610646583762
Change-Id: I33670b943a41611d2c058a42171bca77a22b0d62