[roll] Roll fuchsia [build][repro] Run, backup, run, compare
Previously, we renamed declared outputs in the re-run of the command.
However, it turns out that some actions outputs' contents are sensitive
to output-path/name differences on the command line.
So, we leave the command-line alone, and use the renamed location
to make a backup copy of declared outputs before rerunning.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498352
Original-Revision: 788e7e4c0cb79d07a03c927d98eccae31488a575
GitOrigin-RevId: 9fe25a28086b00c4c1becd5fd7ce23afc230487f
Change-Id: I23d0cb8c9d188373849a3b68365e3741fc0cefc4