[roll] Roll fuchsia [zbi] Do not rely on item type to determine BOOTFS-ness
Today we check whether the type is ZBI_TYPE_STORAGE_BOOTFS in
determining whether to display an item as having a BOOTFS payload. We
will soon invalidate the assumption that BOOTFS-ness can be determined
solely by looking at type* and it also sensible for `zbi` to not make
assertions on BOOTFS-ness that extend beyond the ZBI spec (least of all
in these diagnostic modes).

*We will introduce support to pack ZBI_TYPE_STORAGE_KERNEL items in the
BOOTFS format, but it is not necessarily true that all such items will
be in that format.

Original-Bug: 68585
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499889
Original-Revision: 5287d6d38a3a454e595f4c89afe4fbab214bded5
GitOrigin-RevId: 858edcdba21a10fe1f25dfeb5409ea4fe771c6d3
Change-Id: Ia92d3f35dd2a5b2136236e15d23be17669e72629