[roll] Roll fuchsia [gigaboot] add libavb to build
Adds libavb to the device-side build and ports the necessary sysdeps.

libavb was designed to be forked and modified so this is a little
awkward, but better for our purposes so we don't have to create a second
copy of the code in-tree. Luckily we already have a lot of Posix APIs
implemented so this just needed to add a few more.

In the longer-term, we'll depend on the full zircon_boot library instead
of libavb alone, but this unblocks channel-in-vbmeta work until we're
ready for that.

Original-Bug: 71571
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498358
Original-Revision: abc1ea33d874adae69790de780caf9a6ac10c2d5
GitOrigin-RevId: af1dd85287d7fecefc42784baf2461bc9254b633
Change-Id: I8de52eb91d9c9e912c794471cc14cac3da9d4fe5