[roll] Roll fuchsia [rtc] Cease maintaining the Kernel UTC clock.
As per RFC-0008 we're in the process of removing the original
`zx_clock_get` and `zx_clock_adjust` syscalls along with the kernel UTC
clock.

kcounters show that we've successfully removed the last usages of the
kernel UTC clock and so this CL stops setting the kernel UTC clock. The
"fallback" rtc driver is now no longer doing anything useful and will be
removed in a future CL.

Original-Bug: 31359
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495159
Original-Revision: c66559a933951145cc596abe1d69431a00b277a7
GitOrigin-RevId: 092122124a6eb028f89088b863c429e9fbf8da1f
Change-Id: I704c27f14ffcf6d5d7e625ebc10e627786dac159