[roll] Roll fuchsia [zbi] Support packing STORAGE_KERNEL items
Like with STORAGE_BOOTFS items, STORAGE_KERNEL items are stored in
the BOOTFS format. We introduce support to pack the former using the
same command-line ergonomics as the latter. We do this in part by
introducing a new `--filestype` flag to allow for the selection of one
of the BOOTFS types.

This is an reland of
https://fuchsia-review.googlesource.com/c/fuchsia/+/498698, which had a
buggy assumption in the display logic. That logic has since been
refactored in a parent change and this change differs from the original
only in no longer touching that logic.

Original-Bug: 68585
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499890
Original-Revision: 8c06574ca67a4429d6e1a36571571d4034d17db4
GitOrigin-RevId: 73923b05cd62a68a1e296db0b2d018a5c1eec254
Change-Id: I7c3c7f3cb23b89e5dc29ba16c5f44f4f68c5858c