[roll] Roll fuchsia [pretty] Fix `format_size` rounding in edge cases.
The current `format_size` algorithm will incorrectly round in some
cases. For example, `format_size` emits the string "10.5k" for the
input 10700, but:

  10700 / 1024 == 10.44921875

which should be "10.4k".

The problem comes from two sources:

  * When rounding is performed, the variable `r` contains the remainder
    for the last division by 1024. The rounding code performs
    calculations assuming that the last division was by 1000, however.

  * The rounding code only has information about the most recent
    division. Unfortunately, because the output is in base 10 but the
    units are base 1024, details early on in the calculation can
    determine what final rounding we need to do.

    For example, 0xff33333333333333 needs to round down to "15.9E", but
    0xff33333333333334 should round up to "16.0E".

We update the code to calculate a divisor (a power of 1024) and use this
to calculate the final value.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499738
Original-Revision: 0b55179ad413f3324e11f89c2716112b7a1af74a
GitOrigin-RevId: caba0daa9698efa93b83928bec7d369cd2761070
Change-Id: I268999292f3f719b83f039d56d512f8be844af40