[roll] Roll fuchsia [vulkan][gtest] Add back Vulkan test macro
This macro was originally removed as it uses gtest internal methods
that prevented a roll. Now that the roll is complete, this adds back
the macro, updated for the new methods. Note that this still relies on
gtest-internal methods.

This change requires the new version of gtest, rolled in tqr/340050.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499387
Original-Revision: 89601b040b3a165e356fed2ffc2312096cbc691d
GitOrigin-RevId: 72b62c31f4f611805d246e435f46ebf7b665df49
Change-Id: I499b074864e8c793638b2eae15985a180c5c1b09