[roll] Roll fuchsia [logs] Root archivist logs to debuglog to avoid losing its own logs.
When the root Archivist crashes in the field we currently don't get
any logs from it for the current boot log. If there's a subsequent
crash we'll get them in the previous boot logs if we're lucky, but it's
much safer for the logging service to send its own error logs straight
to the kernel.

Original-Fixed: 69559
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499886
Original-Revision: 2161920b2bb727a56d3cdda8f3c51b5bc92310b0
GitOrigin-RevId: c635f9234f8e94a4e188d1422f9a63c44c3295a8
Change-Id: I6e4e645c1bb173af2d5ec76c35ec91b119049b71