[roll] Roll fuchsia [system-updater] enforce the OTA backstop
In this change, we put it all together and enforce the OTA backstop.

Test: fx test -o system-updater-tests
Test: fx test -o system-updater-integration-tests
Test: manual test by OTA to epoch 1, and verify downgrade OTA to epoch 0
will fail with unsupported downgrade. Note this will not print out with
`fx ota` because it's not exposed in the fuchsia.update state...but it
prints in the logs. fxbug.dev/72118 proposes DX improvements for this.

Original-Bug: 68408

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499981
Original-Revision: 202c37fa01f75c431f61ca824b4d2f7c2ec82178
GitOrigin-RevId: 32037653485204f969eda80c3bbf5c54aa49e23f
Change-Id: I68281056baac22207fb556dbbaf14df52c8fb602