[roll] Roll fuchsia [sl4f] Split trace processing test data out into json files
The large dart file was impossible to view in code search and code
review, and since many of the traces were a single line they were liable
to cause problem for grep tools, with no particularly easy way to
exclude the file.

This change splits the traces out to pretty-printed json files, which
should allow tools to process diffs, and since they are json files they
can be excluded e.g. with ripgrep's -T option.

Cpu and Gpu metrics needed to be modified to have an explicit conversion
to double when reading argument values, to fix 'int is not a subtype
of double' errors.

Fixes: 71931
TEST: fx test sl4f_client_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500078
Original-Revision: 0039c997bbab0c0af006b4efab5cca3cce1aaae7
GitOrigin-RevId: 35da220b69d75e0784a27aedd87316df29367f8b
Change-Id: Ia4c14a5787ce3ec9136052901b9895bda120d65f