[roll] Roll fuchsia [camera] Account for legacy stream disconnect
In DeviceImpl, the table stream_request_sent_to_controller_[] was
used to track whether a "preceding" stream was connected to first
in order to properly configure the graph for the camera pipeline.
This need for an ordered stream start up is a controller bug that
is difficult to fix.

The work around is to ensure that all streams were started in the
right order. The problem with the prior implementation was that a
stream shutdown/disconnect wasn't accounted for in this table.

The fix is to invalidate the associated stream entry for streams
that shutdown.

There is a more controversial part of this fix, which is to try
to reduce the likelihood of timeout portion of the re-order logic
by increasing the timeout significantly. This is not a good fix
for use cases where not all of the camera streams are in use.

Original-Bug: b/181345355
Test: Verify that the "Lost connection to camera" error no longer
      shows up.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498717
Original-Revision: 3423cc59ef5877582c6edfadf0c7511516dd645a
GitOrigin-RevId: a9ddb662787d89e812cf529b146628425d6b674d
Change-Id: I237f6e45091887c9d2a5eb983b80751afe350c0f