[roll] Roll fuchsia [checklicenses] Match file list fix.
Match.File should contain a list of files that that specific Match instance matches.

If it's a project-level license, we don't want to also include all of the project files in that list, otherwise it becomes difficult to verify correctness later. I've created a second list in the Match struct to hold those files (though tbh right now it's not being used for anything, so we may be able to get rid of it entirely)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500799
Original-Revision: 156855166eea21b1295b3b550574f17039d310d4
GitOrigin-RevId: d9a7d18e78fc062868281a0d5a3e39cd2c5a702c
Change-Id: If98d765ffbd2b139f2e72aae29895e3bc26229ad