[roll] Roll fuchsia [tefmocheck] split swarming output into per-test files
This is a prerequisite for associating check failures with specific
tests. See attached bug for details.

And add those files to the test's OutputFiles in the summary. This will
expose the files to ResultDB, Sponge, and the recipes, which may be
useful.

Original-Bug: 67640
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499558
Original-Revision: 3050ffe593b52abb6b5d60f4a75e69bc197228fd
GitOrigin-RevId: 089244c190b43dc1dae7f9542405e9d6a32433bd
Change-Id: Ibded49eee751cfea33941d1774982783335aa500