[roll] Roll fuchsia [build] Remove variant_target parameter from toolchain_utils_action().
This parameter is obsolete and completely ignored. It's a memory of
the previous ZN build, which didn't implement variant builds in the
same way as the Fuchsia build. So remove all the places where
it was previously defined. This should not change the build.

Original-Bug: None
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499938
Original-Revision: 737dd45d7674dfa8941a865710d2ac2c72aa336c
GitOrigin-RevId: e3d3389e654f421d68a2e6a2828128cb0724dcea
Change-Id: Id9731904f76310bd8afab36a35467db89612418e