[roll] Roll infra/recipes [build] Consolidate archive filtering logic
There's now only one chunk of code that filters archives, making it
easier to figure out which archives are included in the build.

Also modify the test data to make it clear that consumers of this
property will need to handle archives with the same name but different
types.

Original-Bug: 67861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/500883
Original-Revision: 5fc62d47c83fc6c1552038ee4342202a4432ec42
GitOrigin-RevId: 37559324287fdec6f797a40e06b0c6dfcab243b5
Change-Id: Ifdb03aa4fc3933706620b4d5102371870610b1e9