[roll] Roll fuchsia [hoist] add facility to disable autoconnect
The ffx frontend is in control of the ascendd host, and wants to
restart it on upgrade. In order to assert when the node goes away
(promptly), the frontend needs awareness of the unix socket link,
thus it wants to take control. This API enables it to do so.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496448
Original-Revision: a587c9340591cd6fed1a5418852092f80fba925d
GitOrigin-RevId: d9e44cb3264b86b56577673c6a2053afddad102e
Change-Id: I9c5ad2008644db71545a6ab59461a8a4445cf651