[roll] Roll fuchsia [audio_core][tests] Show all results, update baseline
Update (tighten up) the saved fidelity baseline thresholds to
reflect the state of the current system.

By default, fidelity results are not displayed at all unless the
full-spectrum of frequencies is tested, so if this is NOT the
case, but '--recap' IS specified, it is reasonable to display
all results (not just a subset) -- but only for the 4 summary
frequencies, not all 52.

Test: fx test -o audio_fidelity_tests -- --recap

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499887
Original-Revision: bdaba1ff38dded28ae5a6b5e39bafbc9fc5c612d
GitOrigin-RevId: 33d106b0edbdf60573df212b4100fb4e1c1e884d
Change-Id: I02dc99aabdc722408f12ba9ed88683aaa39153d3