[roll] Roll infra/recipes [build] Cache compdb after reading
So we don't read it twice unnecessarily when `filtered_compdb()` is
called twice with different filters. This also makes it possible to use
the `@cached_property` decorator instead of `@memoize`.

The problem with `@memoize` is that it stores return values in a dict
keyed by function arguments, so all arguments (including `self` and all
of its attributes) need to be hashable. Python protobuf objects are not
hashable, so `@memoize` prevents us from attaching protobuf objects to
the `_GNResults` class, which is something I'd like to do in a later CL.

Original-Bug: 67861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/500783
Original-Revision: 0920b6987d78c2f562e6cfec3a129c1e53186c4a
GitOrigin-RevId: 275510cb0994a9c9a0e7fb41facd4fc516b5649c
Change-Id: I43c27798252cd731c9c9431312a196a4cfa23ec1