[roll] Roll fuchsia [system-updater] s/current/src
The system-updater convention is to use "source" to refer to the current
OS, and "target" to refer to the OS we are updating to. In this change,
we rename "current" epoch to be "source" to align with this convention.

Test: fx test -o system-updater-tests
Test: fx test -o system-updater-integration-tests
Original-Bug: 68408
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500229
Original-Revision: ff0df64f9ae531ad23fb871d8ed7e1c717b3d6a3
GitOrigin-RevId: a64e5d3386c0ea227c1cb42623db2618abdb789d
Change-Id: Ica1b8d2059f25e0b0f54f309b442d0736f800e90