[roll] Roll fuchsia [tefmocheck] Handles when no summary.json exist
The recipe skips the -test-summary-json flag when summary.json doesn't
exist on disk. This caused a panic in case of test failure without a
summary.json. Silently skip in this case, as the test had failed anyway.

It would be preferable to test this part but this requires a bit of
refactoring to create the equivalent of a mainImpl(). This can be done
as a follow up.

Original-Bug: 72235
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501718
Original-Revision: d99207e8df7a6fe1f221d008f3ff5a5130f413f2
GitOrigin-RevId: f6798a4961cb11a9a941a74d237a58c74c3af8df
Change-Id: Ia2bd5e12fd81c65b05d614eda02a2d4fa9c46e5f