[roll] Roll fuchsia [smc] Add SMC return values to overrun warning
The contents of the x0-x3 after an SMC instruction will contain the
return parameters of the call. For yielding calls, the values will
commonly contain the reason for why the SMC returned. This could be that
the yielding call has completed, or it could indicate that we're
returning to process an interrupt. This CL adds the x0-x3 values to the
debug log statement that occurs when we detect the SMC call overrunning
the preemption timer.

Original-Bug: 69037
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500961
Original-Revision: ac9d7737cb2e1abe6281c2c49ac2d13557d0c09b
GitOrigin-RevId: 842c119cdf30ed3e5c3e296c3174e839c1ef9b0e
Change-Id: I8c15a962e7d4be711595490809bfe950e7783684