[roll] Roll fuchsia [fidlc] Enable syntax version declaration.
In cases where the --allow_new_syntax fidlc flag is set, the parser uses
the "deprecated_syntax" flag at the top of the file to determine which
syntax to use.  If the flag is not set, the parser defaults to using the
old syntax, and errors whenever a "deprecated_syntax" token is detected.

Original-Bug: 65978
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/492578
Original-Revision: 1e14dee06ee61b59adad849065dfff615f8ef005
GitOrigin-RevId: 5dcc4a20e0bffc6b0f8174d17d578d628cf2e531
Change-Id: I2a4489accc286771dbc44165869a72cc2e9bb79d