[roll] Roll fuchsia [pci] Add PciAckInterrupt protocol method
This new protocol method allows a device driver to re-enable their
device's legacy interrupt before waiting. This is done so that the Bus
driver can disable a device's interrupt generation while an interrupt is
being serviced to prevent the IRQ worker being woken erronously when a
driver fails to service a request or does not exist for a given device.

This method is only necessary with userspace PCI, but a noop shim has
been added to the kernel PCI driver for migrational purposes.

Original-Bug: 32978, 71188
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472847
Original-Revision: 23c1a7ca392b691635e2e510e866e4a5467fe4de
GitOrigin-RevId: d706931ac975b4f090d8bc7e316fb2690112922b
Change-Id: I570a2d4cbee847f9bac103a3610c06d5e7c46bfb