[roll] Roll fuchsia [debugger] Work around Clang structured bindings symbol bug.
Clang structured bindings symbols are incorrect as described in the new
comment and in
  https://bugs.llvm.org/show_bug.cgi?id=49565

This workaround shows all "artificial" variables in the "locals" command
to work around this bug. The "artificial" DWARF tag is mostly just used
for "this" parameters which was already explicitly ignored by this
function, so there is currently no known down side to always showing
these.

This hides unnamed variables in "locals".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500813
Original-Revision: 3fdbce6ad45987bd24d34b43bad55ccc861e1570
GitOrigin-RevId: b4834d35f39d706fa77229713ffb0cda86f4ad45
Change-Id: I29e21922b22763c38c889fbf7f15c5ef853ff027