[roll] Roll fuchsia [bootsvc] Use valid handle for `zx_debuglog_create`.
The `zx_debuglog_create` is a (or, _should be_) a privileged syscall,
allowing access to the kernel's debug log. Long term, the goal is that
only processes with the root resource (or a similarly debuglog-specific
resource) should be able to create a debuglog object.

As a backwards compatibility concession, `zx_debuglog_create` currently
allows an invalid handle to be passed into `zx_debuglog_create`.

To make progress towards the goal of removing this concession, this
CL updates bootsvc to pass the root resource to the `zx_debuglog_create`
syscall instead of an invalid handle.

Original-Bug: 32803
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501441
Original-Revision: 89dd0eb1dd8c173315a38b62c2777dd172db36f7
GitOrigin-RevId: 1f272abcb25151a8fa3563ed6bac2f42fedaa999
Change-Id: I677b9dc993b6d111e16f429896bff75ce38f316e