[roll] Roll fuchsia [ffx] improve edge condition behavior on upgrade
The upgrade path detects a new binary version by the hash of the
binary itself. The code now has explicit handling for the underlying
unix socket as a rendezvous so as to not prematurely connect to the
old daemon leading ot failures, and to improve error reporting if a
daemon exits preamturely. A race condition exists where once the
frontend promptly starts to rediscover the daemon it may still
discover the old daemon on the Overnet mesh, as the node still exists
in the peer table even though there are presently no active known
routes to it. The old daemon nodeid is now explicitly ignored in the
post-upgrade daemon discovery.

Original-Fixed: 70486
Original-Fixed: 57264
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496968
Original-Revision: dc6bfac1dfc2c70c6f46c2ff8966968e408168c5
GitOrigin-RevId: 6db57ffee5b71e94fc6f26cdb4ede0432283102c
Change-Id: If0350c7d6c23879f56c941dc8974360685611847