[roll] Roll fuchsia [llcpp] Remove iovec from LLCPP
The existing iovec encode path assumes that the message will
be all bytes or all iovecs pointing directly to input objects.
The new plan is to use a hybrid approach, where some objects
are outputted as direct iovec references and some are copied
to a buffer.

This CL removes the old style of using iovecs to make way for
the new style.

Original-Bug: 66977
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498981
Original-Revision: ca1abfa50150e9f7eb2515bef85197f6b8037320
GitOrigin-RevId: 5dadc5a4414ea6281746b465d659e1f21bcbb072
Change-Id: Ic629b62317503233c4daeafd0308d52535a7ecc8