[roll] Roll fuchsia [tefmocheck] rename per-test swarming output files
ResultDB shows the file base name in the UI, so we want to keep the
base name the same as the unified swarming output file so that it's
clear what it contains.

Also we don't need both the test name and the test index in the
directory hierarchy.

Also simplify a test by removing a loop since there was only one test
struct in the slice being looped over.

Original-Bug: 67640
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502053
Original-Revision: d7d50c8fb978e21bdc577da2c85483bcc2ec4444
GitOrigin-RevId: c2243d32776529c576348e8b901bf7aea64bb7d7
Change-Id: I0ed607ce8a02fe18f9e0f114330b62e6c89a7ab1