[roll] Roll fuchsia [fuchsia-pkg] Use matches::assert_matches
This change removes the custom assert_matches macro, replacing it with
matches::assert_matches and adjusting uses of the macro to specify
additional assertions in the pattern instead of match body.

Later changes will move more functionality to this crate, and those tests
expect this form of the assert_matches macro.  While it isn't impossible
to use multiple macros with the same name, it'll be more clear if
assert_matches always refers to the same macro.

Original-Bug: 64627
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501952
Original-Revision: 153b421d561a069996c4e2e3082b28d907507313
GitOrigin-RevId: f8c83bdedd3c27b637279b48cc040ce4c23303e3
Change-Id: Ie7218ed5fe65d11e08971f9eb4bfeb2aea85f804