[roll] Roll fuchsia [bt][bt-gap] Suppress proptest warnings
Configure proptest to use MapFailurePersistence so that it does not
attempt to write to disk nor walk the source code directory paths.
FileFailurePersistence::Off is insufficient as it still attempts the
directory walk; This issue should be raised and fixed in upstream.

Test: Run the tests and observe no warning output from proptest:
  $ fx test -vo bt-gap-unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500808
Original-Revision: 6d6cba6549909f46b797368d5837a249aacfef6d
GitOrigin-RevId: f12a8c2b4229b8c6a3da33436aa25fd70545f3d2
Change-Id: I195bca0fe5e842eca9d91c3dace1669a54d22a30