[roll] Roll fuchsia [libdoc] accept backslahes in strings
The lexer was interpreting backslahes in strings. It would have been
correct for a computer language but it is not correct for documentation.

This CL removes the code which used to interpret backslashes and add
tests to check that we can have a backslash at the end of a string.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499578
Original-Revision: 5565ea8fdaad912a14cb7d07faecdd71bfbfd12b
GitOrigin-RevId: ad95aad925e8a48b9857f48b7cb30c29685f49f2
Change-Id: Ib7ccaec129da2b24480987ade618c2520ce654f8