[roll] Roll infra/recipes [ninja] Drastically simplify ninja_wrapper.py
The vast majority of the ninja_wrapper.py logic is dedicated to parsing
the Ninja stdout in order to get a detailed JSON summary of which tests
failed. However, we don't consume that JSON file anywhere.

In fact, we weren't even retrieving it from the ninja_wrapper.py step
because it was only added as a step log when the command failed
(add_output_log="on_failure") but we were setting `ok_ret="any"` for
the Ninja step which caused the log to not be added even on failure.

Now the script's only responsibility is extracting a concise error
message from the Ninja stdout, which may be hundreds of thousands of
lines long.

If we ever need to restore the old logic it shouldn't be hard to do it
using git history.

Test: https://ci.chromium.org/swarming/task/5250e93d86bdac10

Original-Bug: 67861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/502138
Original-Revision: b018bf44891f05f166af09aa89c3a342e935f3a0
GitOrigin-RevId: b1c39f9b38d13741fdc1d86e6e283ad639806297
Change-Id: I7a88e4b220c6f78940c8725ef3698cc808d9f097