[roll] Roll fuchsia [flutter] Use fake display for tests
The test was hardcoding 200x200 for the child view and overlay logical
sizes.  We use a default logical->physical scale of 1.7, so the overlay
gets cut off if the display is < 680px in x or y, failing the test.

The solution was to use FractionallySizedBox() to stop hardcoding the
child view and overlay sizes.

BUG: 70433
BUG: 67316
TEST: Deflaking flutter-embedder-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502231
Original-Revision: 5cd7eeae4ead3e7a2a77cb2763c3d60af9e21d57
GitOrigin-RevId: 7bf8b31b9ebc3ca9b451b86c6aba2a14f0ba52e1
Change-Id: Ifeecf48cb9ad3c5a56b0a29ac9bbdc6f71df3e04