[roll] Roll fuchsia [fxfs] Device, Buffer and BufferAllocator
Device is the interface for reading and writing from a block device.

Buffer is a sliceable, contiguous buffer which is taken out of a special
memory region (for example, from a mapped VMO) which is appropriate for
transferring data to and from a block device. Buffers are owned by the
code which requests them, and automatically clean themselves up when
going out of scope.

BufferAllocator is the interface for creating Buffers which can be used
to transfer data to and from a Device. Generally, clients will simply
ask the Device for a reference to its allocator, and allocate blocks out
of that.

Each Device implementation is expected to implement its own BufferSource
which is the backing pool for Buffers. For example, a Device which is
backed by a real block FIFO would allocate Buffer objects out of a VMO
which is attached to the block device (i.e. has a vmoid). A later CL
will add such a Device.

Tested: Unit.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500279
Original-Revision: 4e69bc6fec186d8f85bc4c57a72291f134c233b5
GitOrigin-RevId: af7dd895fb30ef642d4798134aee8a5c248ccdb3
Change-Id: Ie8274c79df72f8da6f14a59cb78a9e0c10fddf3a