[roll] Roll fuchsia [kernel][lockup] Fix DAP powered backtraces
Fix the backtraces during fatal condition violations detected by the
lockup detector that we produce using register dumps harvested by the
DAP.  The problem was that the shadow call stack register (x18) is
always pointing to the _next_ place to stash a return address, not at
the last address stashed.  In addition, the link register (x30) is
what holds the immediate return address of the current frame at all
times.

Because of this, frame 0 of the backtrace was correct, as were frames
2-N, but frame 1 was showing some stale data on the shadow stack,
while LR held the real value.

Fix this issue by appropriately dumping the LR as frame one, and
attempting to back up in the stack before producing frames 2 and
onward.

Original-Bug: 70724
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502711
Original-Revision: 484b18ef5e15ed2b48e37507226d5202ce70489b
GitOrigin-RevId: 7e8da16e23e4d458587c2732b466594be5edc848
Change-Id: Ie316de8780053444ffae37db5401e87b632341cd