[roll] Roll fuchsia [media][tools] Extend rates, improve confidence
This CL builds on the recent extention of downsampling ratios,
by changing the profiled "extreme" ratios from 8:192 and 192:8
to 1:192 and 192:1. The 192:192 ratio is added.

To improve confidence intervals for profiler measurements, the
default number of iterations is increased from 500 to 1000.

Execution times are slightly increased:
* For the perftest subset, from 6 sec to 8 sec
* For the full developer set, from 180 sec to 245 sec

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501244
Original-Revision: 2ea8c12e4860168481844c0ba67af749beddeebb
GitOrigin-RevId: 19dbd7cc5ebab60a0a83f3e078b78498f9dce362
Change-Id: Ia01861b38e7b4434440f1cd13ed307e99a9d0455