[roll] Roll fuchsia [wlan][brcmfmac] Add the MLME_LINK_FAILED reason code
A special reason code MLME_LINK_FAILED is added to
fuchsia.wlan.ieee80211.ReasonCode so the MLME can indicate to SME in
the reason code when the link between the client and AP fails for some
reason other than receipt of a deauthentication or disassociation
frame. The most common cause of this is loss of signal with the AP.

The impact of this change is that metrics emitted by Fuchsia will now
more clearly show which disconnects are being caused by MLME link
failures and which are caused by deauthentication or disassociation
frames.

Original-Fixed: 72469
Test: Connected to mobile hotspot and walked away from the device. The
new reason code was reported by brcmfmac as expected.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502800
Original-Revision: 5008d1464e89d68b7eab576b2f63269f27a886fa
GitOrigin-RevId: 0457ef5c4b86154ff55b51421fe32b2a316b9c27
Change-Id: Ied5870544c168e34ddfe415ce5348652dd9b770a