[roll] Roll fuchsia [stress_tests] Store last operation in FVM stress tests
FVM stress test operations are not atomic. It is possible
for an operation to complete partially before a crash occurs.

If the volume actor fails to perform an operation because
of a volume disconnect, store the operation so that it can
be performed again when the volume is reconnected.

Original-Fixed: 70180
Test: fx test fvm-crash
      fx test fvm-nocrash
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484100
Original-Revision: ac8f217aae4f0e71996ff9a67796b98eb993ec8c
GitOrigin-RevId: 4a6cfb1da51e3b48c574449294a4595289fb32a7
Change-Id: Ie0c39f3b32466bf0e57ecd69cc1bf5881d6b888c