[roll] Roll fuchsia [block_client] Convert RemoteBlockDevice to trait
Turns `struct RemoteBlockDevice` into `trait BlockClient` and
`struct RemoteBlockClient`.

This is pre-work to create an in-process fake implementation of
BlockClient which can be used for in-process tests. (Currently clients
have to stand up a remote ramdisk to test against this class, which is
heavyweight.)

Leaves RemoteBlockDevice aliased to facilitate a soft migration.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502918
Original-Revision: 180bc38bd4c0c23fc4eb9abdccba4ba52473e55c
GitOrigin-RevId: dacb26223698dd596d5da79c259b020f0136084c
Change-Id: I46c5503dbe35848f6270998c632683987a235fb4