[roll] Roll fuchsia [zircon][scripts] Only rebuild the JSON file when generating docs.
`update-docs-from-fidl` pulls information from a generated JSON file.
If the output JSON file does not exist or is out of date, the script
will perform an `fx build`.

Because the input VDSO files are deep in the build dependencies, each
change to the input file typically requires recompilin a large amount of
the system. This makes iterating on documentation painfully slow.

Update the script to only build the JSON file, and not the entire
system.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501439
Original-Revision: 3869f71959cd04a25198004c3864261d859f4b01
GitOrigin-RevId: 2f931e0ddde72663b5f6ee877fac18831c2ee32e
Change-Id: Ic40ed6fcf97a6f3e8fee8a529df0dd41d7ba571b