[roll] Roll fuchsia [media/vnext] update fidl libraries for media vnext prototyping
This CL updates the fidl libraries to //src/media/vnext/fidl. These
definitions are a 'work in progress', but are checked in so that
members of the media team can write code against them to refine and
expand on them. The intention is that they will eventually move into
//sdk/fidl and become part of the sdk.

Definitions in fuchsia.media2 and fuchsia.mediastreams are based on
design documents currently under review. Definitions in fuchsia.audio
and fuchsia.mem2 have not yet been subject to review and are
speculative. fuchsia.mem2.BufferProvider, in particular, is a 'toy'
alternative to sysmem and it missing many features.

Test: no tests - fidl definitions only
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/503581
Original-Revision: b652eef506c86c2c7fd827ab5a9b188fc267d8d0
GitOrigin-RevId: 848cb28087d8e9a6678e42d9a209d3eca3de2a4e
Change-Id: Id822ea554f0d28779b868939bf2b7dfec1ecefed