[roll] Roll fuchsia [lib/arch][x86] Introduce X86MsrBase
This is a convenience class to inherit from in defining an MSR register
type. It saves on boilerplate (especially the eyesore of a static cast
of the associated X86Msr enum value back to uint32_t) and mirrors the
base types we have introduced for defining system and CPUID registers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501662
Original-Revision: 0e66ea12c93d6bd321b2909e0a8787cdbce51f45
GitOrigin-RevId: 2b3f46cfaa0223f8c82fa5c7f2d9013685f96234
Change-Id: I688fd58a2b6840a489432a394900101c2e7eddd0