[roll] Roll fuchsia [ffx] plumb ssh port from target to host pipe
An optional field is added to the target that enables us to override
the ssh port on a per target basis. This port is plumbed through to
the host-pipe connection setup, but is not yet plumbed upward to
target add, which comes next.

Original-Bug: 66457
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487857
Original-Revision: c31fb037fb826683cdbfeaa57957a2a9fd15bde7
GitOrigin-RevId: ae548af22fb1144550e7e7e2459702de89a6a6fd
Change-Id: Ic5b204fae1f0f8a9985244ab6c359a5272a47a49