[roll] Roll fuchsia [flutter-embedder-test] Moves the injected services into a function.
Removes the use of a fixed-size array in injected services
initialization.

Needing to keep a manual count of the number of elements in the array is
error prone.  And, the use of the service list seems simple enough not
to really require the array optimization.

So, I propose to remove the size constraint, and replace with
std::vector.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/503780
Original-Revision: bb527537874d72006ae11e516306c901e0381abb
GitOrigin-RevId: 5887a17127fba72c2b41d2853d2d570eb3bee4f5
Change-Id: I1bd41fb2b3ff722eb57fad9c563768515a9457f5