[roll] Roll fuchsia [scripts][fuzzing] Improve handling of non-directory corpus targets
In several places, we had implicitly assumed that GN corpus targets
always directly correspond to a directory in the source tree containing
corpus elements. This assumption does not hold true for corpus targets
of the form `//a:b`, where `b`≠`a`, so we did not handle these cases
correctly.

This CL modifies the way `fx fuzz` maps GN labels to package namespace
paths to match the approach taken in `fuzzer_test_component`. It also
ensures `fx fuzz update` gives an explicit error when run for fuzzers
with a non-directory corpus target, as automatic buildfile generation is
not possible in this case.

There remains a potential for name collisions between targets of the
form `//a:b` and `//a/b:b`, but this is not addressed in this CL.

Incidentally, the copyright year for generated corpus buildfiles has
been freed from being stuck in 2020. May this bode well for us all.

Original-Bug: 71634
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499878
Original-Revision: 0c1babbfe4698a478f04f5ed3c506ed87cadb00b
GitOrigin-RevId: c2403fba468529cd55cc89810fcb5ed7fea9988d
Change-Id: Ie78b457d26bc124b2bb26429543e08285f33444c