[roll] Roll fuchsia [bt][hfp] Add AgUpdate enum.
The HFP component deals with procedure updates at the AgUpdate level
rather than at the individual at::Response level. This allows the logic
to convert an AgUpdate to a set of at::Responses to be kept to
a single layer of abstraction.

Test: Unit tests cover all procedure conversions
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/503840
Original-Revision: 9ec6cd5d6bf0a43f664eba25dc422b980fafdc06
GitOrigin-RevId: 71170d633cb2f3b26141c3c4228b1a2dd46c40b2
Change-Id: Ie43e25e95b20364a4355a78b43d428bdcbbe1481