[roll] Roll infra/recipes [goma_client] Fix checkout logic when build input != repo
This change fixes the behavior of the recipe when the build input does
not match the repo we're trying to checkout. This will happen in the
case of goma_client builds which are manually triggered through LUCI
scheduler UI.

See go/bbid/8852399315497283600 for example failure.

Original-Bug: 71887
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/504358
Original-Revision: dbed82672c16334ba7d236ec2aff71d695b8719d
GitOrigin-RevId: c6dbb540c509711397b4fe38c48ea68b7dbee649
Change-Id: If05a59223198962a437d045d61dc7f243017404f