[roll] Roll infra/recipes [checkout] Use `-rebase_branch` for CQ.
Jiri takes a `-rebase-revision` flag to `jiri patch`, we had an unused
`rebase_revision` arg piped around in recipes code that was never used.
This replaces it with a new flag that will rebase on an arbitrary branch
instead of "master" in the absence of a `-rebase_revision`.

Original-Bug: 69324
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/501326
Original-Revision: 59726e824e8994d949c04be87aed8f414e6a8dd2
GitOrigin-RevId: 6a4fdfef05c7d47a7ea720dcc10a5ff4d3beb9e2
Change-Id: Id084ec2f163cb9b8b0d1d53eed1651c5471b587b