[roll] Roll fuchsia [fidlgen_cpp] Deduplicated request/response fields
This CL fixes the "// TODO: Consider factoring out common
fields between Request and Response." in protocol.go

It introduces a "Message" element in the IR, which represents either a
request or a response.

The existing Request/Response field (an array of parameters) are rename
to RequestArgs and ResponseArgs respectively.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501504
Original-Revision: bb72c698515d10fa9901db77b2ab48f1768271f0
GitOrigin-RevId: f3c173651bc31eebd8aae1748aaca3d1c0191ba6
Change-Id: I25a1e8a6a1fcad37f52ca1ee73f0ef14ff296410