[roll] Roll fuchsia [fx] Add `fx setup-go`
With fxrev.dev/503479, getting vanilla Go tooling and IDE features
working will no longer just be a matter of setting an environment
variable. Instead, we'll need to create a bunch of symlinks so that Go
knows to interpret the root of the checkout as the
'go.fuchsia.dev/fuchsia' module.

This is a complex and error-prone process if done manually, so add a `fx
setup-go` command that creates all the necessary symlinks (as well as a
couple extra go.mod files to prevent a few subdirectories from being
searched for Go packages).

The command also has a "teardown" mode, invoked with the -d flag, under
which it will delete all of the symlinks and go.mod files it created.
This will be useful for getting back to the state of the checkout used
in infra and on other devs' machines, in order to test bootstrap.sh and
other scripts that need to use the Go toolchain but can't assume that
`fx setup-go` has been run.

Original-Bug: 35565, 67547
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504178
Original-Revision: 549addc7e39a41514e1941257f66f7b72232160f
GitOrigin-RevId: 8465a22dabeab0f86f481dd20c37bdda6ebf9b1c
Change-Id: I14c59dd0f5578856ef4c03ae94640633778e23e0