[roll] Roll fuchsia [wlan][policy] fix mixed security recording
This changes the security type returned from network selection to
reflect the security type of the saved network config, not of the actual
scan type.
The scan security type is used to match loaded saved networks with scan
results, and the actual scan security type is not used after network
selection, but the config security type is needed to record whether the
connection is successful. Past connect failures are used to choose when
there are multiple options for connecting.

Original-Bug: 70964
Test: modified unit tests to reflect the security type that network
      selection should return:
	fx test wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490937
Original-Revision: 407393414e673453c28527341fb2e3c5a73098c2
GitOrigin-RevId: b0c35db136de9a9ad6acd0724a24b3e717391929
Change-Id: I268ec2dae89d3076404c0cb00400b603a91217eb