[roll] Roll fuchsia [fidlgen_cpp] Fold LLProp into Method/Protocol struct
We already started computing the allocation strategy in ByteBufferType.
This CL combines that with other memory-allocation-associated fields
into a single Allocation struct.

Doing so revealed a bug where the existing ByteBufferType is the same
for both client/server. In reality, the allocation choice is dependent
on all of client/server context, message flexibility, and message
direction. Four Allocation properties were computed for client/server
and request/response respectively. For example, the server allocation
for sending a flexible response is bounded, but the client allocation
for receiving the same response is unbounded.

In the sychronous event handler, we would now also take into
consideration flexible messages when determining allocation strategies.
This fixes another potential issue where the server could send an event
containing a flexible table with extra fields, which pushes the message
over the calculated size limit on the client side.

TEST: added allocation strategy test using fidlgentest.EndToEndTest

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501506
Original-Revision: 20ba1849399d492a922629e4ebea10b86bd7052c
GitOrigin-RevId: a6be6fba8cfe9a5bbd059a5ee284886426fd2fc2
Change-Id: I0e735d5457b3873c4873ab2209b3a03a13659359