[roll] Roll fuchsia [blobfs] Wire up blob creation to new pager.
Hooks up the blob creation path to the new pager. This avoids creating a
PageWatcher for the blob in the new path and all of the functionality is
now in the Blob object.

This involved creating a callback that handles the differences between
VMO creation between the pager and using that in the blob loader. A nice
side effect is that the pager dependency from the load is no longer
necessary.

To support removing the PageWatcher completely, some additional ifdefed
variables were added to the Blob object. This is unfortunate but makes
this more clear about what's currently happening.

The new paging system still does not net work properly and is off by
default so there should be no behavior change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504238
Original-Revision: 660fe081f0870f11fcd9d9c6864ce6d1f1334243
GitOrigin-RevId: 4d590f50bffc36afb510830ad5eca40d66801e37
Change-Id: I92b0aa9468a9cc7602756e8a8d35d714f2aeddb4