[roll] Roll fuchsia [netstack] Check inspect metadata errors in tests
When reading inspect data, include the metadata.errors field in the
error returned if the payload is empty. This will assist in debugging
cases where an inspect node is expected to be present but is not
present.

Original-Bug: 71452
Multiply: netstack_reachability_integration_test:30
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504661
Original-Revision: e7710a3c50ef63cd340299097254cf6f4dcf89d6
GitOrigin-RevId: 5f2e596c6c1dad2836b98db3db781502fcc5c08f
Change-Id: Id07f776c1c15dd4312a285516c2b35e431c1737b