[roll] Roll fuchsia [build] Suppress lsan for vboot_reference
https://fxrev.dev/462880 created this allow-list of targets with LSAN
suppressed, but missed the target //tools/vboot_reference:cgpt, which
suppresses LSAN errors:
https://fuchsia.googlesource.com/fuchsia/+/master/tools/vboot_reference/BUILD.gn#144.

That allow-list change broke the cgpt target because it was no longer
allowed to suppress LSAN errors.

It seems that that target is not built in CQ, which is why the change
was able to land. However, the target *is* built on the host-prebuilts
builders that build tools and upload them to CIPD, so those builders
have been quietly failing for 3 months:
https://ci.chromium.org/p/fuchsia/builders/global.ci/host_prebuilts-x64-mac/b8852299601903060720

Original-Bug: 72680
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504841
Original-Revision: f76ab9d1e3ae445fea0f20136e898feb3a5571f4
GitOrigin-RevId: cfc416fca1a3f82f1ed31769598f449833ae9a7b
Change-Id: I80554f1b011abc394eb5bc8659748089591c6d99