[roll] Roll fuchsia [tefmocheck] copy output files from tests to checks
When a string_in_log_check fails for a particular test, this will make
it much easier to find the root cause, since the LUCI and Sponge UIs
will show the per-test log which caused the check to fail. And this will
hopefully  help us understand the attached bug, which is currently
a mystery to me.

Original-Bug: 72721
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504939
Original-Revision: c4e03e6469531fb760025e3eb72a415643a34e26
GitOrigin-RevId: 99a0c2531376d8314650eff96cdafe41793f044b
Change-Id: I9287f88299378e975b104ee5dd8be6727951c000