[roll] Roll fuchsia [preflight] Correctly detect M1 chip when running in Rosetta.
When the ffx binary is compiled for x86_64 it will automatically run in
Rosetta - which hides M1 chip detection from the tool. This change
teaches preflight to look for a signal that it is running in an emulated
environment when in x86 mode.

Original-Bug: 70728
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504552
Original-Revision: 6bfa44ce729b4dd766558d99736043d6932e94ba
GitOrigin-RevId: da92d27863eb214b2d18a1434dbb0073f12e801c
Change-Id: Ia0905fdb16cad792700b8d877659a0a3a8c11a9b