[roll] Roll infra/recipes [cherry_pick] Add option to short-circuit after ensuring branch
The cherry_pick recipe already has logic to ensure that the target
project's branch exists. This change adds an option to short-circuit
execution after that step, so that release managers can create branches
without introducing any cherry-picks.

This is useful when we need to bringup a new branch for contributors to
author new changes on, where the initial CL (on main, or another release
branch) should not or cannot be cherry-picked directly to the target
branch.

Original-Bug: 72783
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/504609
Original-Revision: 2a636bbe46d66ab9a1d5a4f643c6d3aae71863b6
GitOrigin-RevId: 3d8b02252c5dc98c62d624597013b81667a9eaa5
Change-Id: Ifefba948a23ada62a49f7d347f192c7cdb0916a8