[roll] Roll fuchsia [blobfs] Implement read-only health checks
Adds a health check to the BlobfsVerifier service which verifies up to
1MiB of data (arbitrarily chosen to give reasonable confidence in health
checks without being too slow) for blobs in the cache.

We only pick blobs in the cache (i.e. those which have been opened since
the system booted) so that we are certain that we do not have false
negatives by verifying blobs that are not actively used. For instance,
if some ephemeral package was corrupt, we would not want it to fail the
health checks after an OTA.

This might mean that in practice few blobs are actually verified (since
we do the verification early in the boot), but false negatives are
significantly worse than false positives here, and at least we would
know that the important early blobs such as the binaries/libraries of
the update stack are valid.

Also did a bit of refactoring in blob.cc (noticed that LoadAndVerifyBlob
was only used for the null blob case).

Original-Bug: 70920
Tested: Unit tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502098
Original-Revision: af0ca1b29ec383da714ae48c58b62c14632b6961
GitOrigin-RevId: 4663c69fed4b50debf77ea4a5926f5bb39c0e509
Change-Id: I5c44cd90a30dbbfd61007e4235c960d4d0120a46