[roll] Roll fuchsia [debugger] Support 0-based line table files.
DWARF defines the 0 file name index in the line table to implicitly
refer to the file name of the compilation unit. Before DWARF v5, Clang
did not emit this and we treated it as an invalid entry.

With DWARF v5 support, Clang started to emit this (even though it was
always valid). We now allow a 0 line table entry to support these
symbols.

Original-Bug: 26814
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504844
Original-Revision: 0d46509c31e33e8f92e435fdee66f3c8116fd1ee
GitOrigin-RevId: 4acd2cb2d072ae68238aef646c9882a6ddb64a13
Change-Id: I477d8fbf7bbf30012c8aa86f29b99f6d90da1f62