[roll] Roll fuchsia [fxfs] Fix skip-list locking
Prior to this CL, the skip list unsafely used a RWLock. This CL
addresses that.  With this implementation, writers cannot block readers.
Writers are sequenced i.e. there can only be one writer at a time and
writers are guaranteed to eventually make progress (so long as readers
do).  The approach taken is similar to lock-free skip-list
implementations.

This CL also changes the iterator API so that seek now returns an
iterator rather than creating an iterator and then having to call seek.
This simplifies iterator implementations which no longer need to keep
track of the special state that exists prior to calling seek.

The merge API is still a little odd, but we can revisit that later.

MULTIPLY: fxfs-tests: 100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504007
Original-Revision: 8157ad6a9088bb6ec96dae496168aee26d380e8d
GitOrigin-RevId: 0be8a27a3878c5e4788ff5043decaa71a0a3ace8
Change-Id: I3a75017ffb32211650229a06cba0d79b7773fe93