[roll] Roll fuchsia [bind] Add true and false statements
The `false` statement replaces the `abort` statement, and the `true`
statement is new. Both of these must be the only statement in a block
(either nested in an if/else, or the only statement in the bind rules.)

The `true` statement is semantically a no-op because a bind rule will
match if it reaches the end of all of its conditions. It is
syntactically useful for writing a bind rule that always matches, or in
rare cases where you might otherwise want to write an empty block in an
if/else statement.

To avoid giving the impression that bind rules are imperative programs,
rather than declarative conditions, these keywords must be the only
statement present in their block. If any statement followed a `true`
statement it might be easy to incorrectly assume that the statement
would be ignored.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/494600
Original-Revision: ab81430aac20e5b224e49ad810ff822d24c9c388
GitOrigin-RevId: 1902b33249df3a35e0b973ced80c8ec2f8c34c67
Change-Id: Ic5a21f9fd5a157793a0cbbd420e334df254fe51f