[roll] Roll fuchsia [fidlgen_cpp] Replace unit types with namespace enum in `Kind`
In order to identify FIDL declarations, we employed marker types and
took advantage of default initialization. This trick has created
confusion and prevented defining an interface like the following:

    type Decl interface {
        Kind() _SomeKind_  // What to put in "_SomeKind_"?
                           // It is polymorphic.
    }

It turns out changing this unit type pattern to use namespaced enums
is just as safe and readable, while allowing us to properly define
the Decl interface.

As part of this change, we store IR value themselves into Decl (and
implement the Decl interface on the IR structs), as opposed to storing a
pointer to the IR values. This has the nice side-effect of avoiding
unnecessary boxing (see
https://golang.org/doc/effective_go#pointers_vs_values). In particular,
the Decls should not be mutated by the templates.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504555
Original-Revision: 9d0f5568351cf19ceb857ddaf2a7e31247ecd8ea
GitOrigin-RevId: e9dc11a1ec3eeaa1f9e3bacca85d98caf2d83c43
Change-Id: I26c6646ce759c65525e5a9ffe88afea91fd290b7