[roll] Roll infra/recipes [fuchsia_roller] Make default ref configurable
Previously, if a roller didn't have a triggering commit (e.g. because it
was triggered manually through LUCI scheduler) it would always roll the
repo to refs/heads/master. However, that doesn't work if the repo's main
branch is not named "master".

Now, the default ref that we use in case of a build with no triggering
commit is configurable via a `default_ref` property. `default_ref`
currently defaults to `refs/heads/master`, but I'll make a follow-up
config change to set it to the appropriate value for all rollers that
track a branch besides master.

Original-Bug: 72787
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/505558
Original-Revision: 2daa0e8c3e5df5f2d503dea7abdf9be4cd890aa6
GitOrigin-RevId: ebfda7cb3ff414023ee169a7e3707a4c741da358
Change-Id: I0f4e3d98eb8c80cd7282d50aa165d240b25e3fdd