[roll] Roll fuchsia [stress_tests] Make CM stress tests hub-based and multi-threaded
The component manager stress tests now use the hub, rather than
maintaining an in-memory representation of the component tree.

Each actor connects to the hub separately, traverses to a random point
in the hierarchy and creates/deletes a random component using the
exposed fuchsia.sys2.Realm protocol.

Errors encountered during traversal, creation and deletion are
silently ignored. The assumption is that the operation could not
complete because of the actions of another actor.

This leads to interesting scenarios where, for example, a component
being bound to in one thread is also being deleted in another.

There are now 10 identical actors that do such operations across the
topology, each on a different thread.

Test: fx test component-manager-stress-tests-alt
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504218
Original-Revision: 387f040cdc369be5917881a2142d59abfd8f5197
GitOrigin-RevId: a4c423bdc4358c439bf0605724298ba52e06c7d7
Change-Id: Ie908396af62293502e2fa3bc6593469458c85e8c