[roll] Roll fuchsia [wlan] Include BYTE in SSID and PSK length constant variable names
SSID is limited to at most 32 bytes and PSK must be 32 bytes. It
is worth including BYTE in the variable names MAX_SSID_BYTE_LEN and
WPA_PSK_BYTE_LEN to make clear that the limit should be applied to
bytes and not characters. In particular, since SSID can sometimes be
interpreted as a UTF-8 string, it's possible for the SSID length limit
to be misinterpreted.

Test: Pure refactor

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501942
Original-Revision: b5f65a30b345fa1fb23b2435372c214da67aa0ea
GitOrigin-RevId: a4132f75d282f77dee3256b7b511f7b81855e87e
Change-Id: I8f3a98ff85d92ffa64988bac907bde06935ce79e