[roll] Roll fuchsia [tefmocheck] simplify splitLogByTest().
In investigating the attached bug, I discovered that the per-test logs
are sometimes incorrect. However I haven't been able to reproduce the
issue locally, so I'm just trying to simplify the code in the hopes of
eliminating bugs. Even if this doesn't eliminate the issue, it seems
like an improvement in readability and performance.

Original-Bug: 72721
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505567
Original-Revision: 09b9f713b622cbb7b4d052060614a085ccd1d3cc
GitOrigin-RevId: f245ff411910759494e3d89d78f5c4c395e63eac
Change-Id: Iae68ea856ae491e3edb570c7f1398c359be68c58