[roll] Roll fuchsia [harvester] Add FIDL support
When harvester is run with the --fidl argument, it will now publish the
new fuchsia.systemmonitor.Harvester protocol. We've also added a new
component manifest to run with this argument, and configured it to
auto-start in sysmgr.

The protocol only has one method, ConnectGrpc, that takes a socket and
establishes a GRPC connection over which the legacy GRPC protocol can be
used.

Original-Fixed: 69454
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500939
Original-Revision: 87e5e65418fca55e240390e6310836dc0896a0a0
GitOrigin-RevId: b167112af24c04e47bb04c7b7569f08994206743
Change-Id: I89d6fb03bd1834380339950ef33dea87da6b3b84