[roll] Roll fuchsia [netemul] Use `RealmOptions.name` for realm moniker
The optional `RealmOptions.name` is combined with two static
monotonically-increasing counters, for the sandbox instance and the
realm instance, to create a unique moniker for the root of the generated
realm.

This CL also introduces the `ManagedRealm.GetMoniker` method to enable
clients to retrieve the component moniker of the root of the runtime-
generated topology that contains the `ManagedRealm`'s constituent
components. This is necessary in order to test that the moniker is
correctly set.

Original-Fixed: 72169
Test: fx test -d netemul-v2-tests --test-filter *realm_name*
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505439
Original-Revision: 81c3f9b521c262aeb78c985f4b5d50e5adee4faf
GitOrigin-RevId: 9ca5422ec58ce353ddc5d3f36ddc226b8d522fec
Change-Id: Ic09c2c672a65b6a38d29231834a0d9066f94f50e