[roll] Roll fuchsia [ffx] Choose shortest tool path when getting a host tool
The SDK provides, typically, three paths for each tool we ship. Two of
them are arch-specific, and one of them is the "default" arch. We
previously just took whichever we happened to find first, but with this
change we'll always take the shortest, which we expect to be the default
one since usually the difference is that the path doesn't contain an
arch specifier.

The word "legacy" appears a lot around this "default" tool, so I suspect
this may break again soon. Looking in to a permanent fix.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505842
Original-Revision: 0d2f6eac2e3be693861f1471faf3459de78f0b27
GitOrigin-RevId: 8f347bbd99baa9c5fc0d899bd1418e117ecf2de8
Change-Id: Ib1dcf0b4bbe6e3d40f11ef582f7d6c3642d3aa64