[roll] Roll fuchsia [time lib] Add a new Transform struct.
This is similar to the ClockTransform in ClockDetails in that it stores
the linear mapping from monotonic time to a synthetic time, but this
struct also stores the error bound and change in error bound over time.

This will be useful in implementing the timekeeper frequency correction
algorithm.

Original-Bug: 56868
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505566
Original-Revision: 9df8421687c3e74e15bbbf7c68f62d78ef0030c3
GitOrigin-RevId: cb8348af3107bb56be5c7ed427b10d8fc409ecae
Change-Id: I8350c10ab9b3192dc24d3742a6a456bdedf5f22f