[roll] Roll fuchsia [fxfs] Add reference types for Buffer
Adds BufferRef and MutableBufferRef, which are arbitrarily sliceable
reference types that are used for transacting with a Device
(read/write).

Reference types are necessary (instead of, e.g. &Buffer and &mut Buffer)
because we need to support slicing buffers in the
ObjectStore::read/write code. A slice of a buffer must necessarily yield
a reference of some kind, and adding a reference type was necessary to
make the lifetimes work out (since something needs to own the reference
and have the lifetimes appropriately scoped).

Buffer references must still be aligned by the time they are used for
transacting with the block device. Assertions are added to ensure that
we catch this at runtime.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506164
Original-Revision: 92f34be61fc62e7374bdebdfed1e263cc97f6eff
GitOrigin-RevId: 4aad5de9bf21f4b1b5fff0f04c1c298469383021
Change-Id: Ia9f74a41adb511c5e2f1b515185d58c0a72af456