[roll] Roll fuchsia [scenic] Allow RootPresenter to be relaunched.
The main bug was that DisplaySwapchain::Flip() was being passed the
display id instead of the layer id.  Consequently, when a new
DisplaySwapchain was created with layer id == 2, it continued to
pass display id == 1 to Flip(), resulting in a black screen.

A secondary issue is that the layer was being destroyed while still
associated to a display.  The display controller logged this as an
error with no other apparent ill effects, but who knows what might
happen on other hardware?

Original-Fixed: 73050
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506786
Original-Revision: f10573a5f6af44b6ed326ca8c71aa028f9368864
GitOrigin-RevId: b67bd0026877307fb5ff0973521c5d2be9ef782c
Change-Id: Ied778961a067debd7cd809e1a33ac2dfbc1ba677