[roll] Roll fuchsia [dev-operation] Migrate off fbl.
We need dev-operation for writing out-of-tree drivers that use usb,
such as the example ftdi driver.

We do not want to include fbl in the SDK, but there is no
std equivalent for the intrusive_double_list. The options proposed
were to downgrade the ftdi driver to use c-style usb requests,
or make a branch of the required fbl libraries. Branching
fbl seems to be the easiest path forward, and since we eventually
expect to deprecate dev-operation, this should not increase tech debt.

I have copied the required fbl headers into lib/operation/helpers,
and changed the namespace to operation. I also changed operation.h's
use of fbl locks to the std locks.

Original-Bug: 71471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506687
Original-Revision: b9f0319e96a10d8940cda4e91616abe87623dafa
GitOrigin-RevId: c5f797fb373e6f653cde73354b8263e8feed3307
Change-Id: Icd0a4085c65bf5b4d7e242cdc9c627862fe4c0f3