[roll] Roll fuchsia [kernel][phys] Ensure the x86 GDT register is correctly aligned.
When the GDT register is loaded on x86, we need to pass in a 10-byte
memory value. Ideally, this should be aligned such that
`alignof(GdtRegister64) % 8 == 6`, because the CPU will load a 2-byte
word followed an 8-byte word. In practice, this doesn't matter too much,
because x86 will happily load unaligned words in most circumstances.

However, the misaligned GdtRegister64 has recently started causing
problems for the linker/relocation code, because the compiler attempts
to emit a pre-constructed GdtRegister64 in the program's .rodata
section, which in turn requires an unaligned reloc.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506621
Original-Revision: 5deb37aa3fd27d5825859dc4ecb41a8069307fed
GitOrigin-RevId: 4799825f302fc2b7bbcfef84eafb9f8d4863ed7d
Change-Id: I945afe6866bae1459ce168b61e8d6681a21c4318