[roll] Roll fuchsia [fit-promise] Add dependencies on fit-promise
This is a new SDK library into which the "promise"-related content of
libfit will be moved. This is the first step in a soft transition.
Currently fit-promise is empty; after everything that should be
depending on it is doing so, the content will be migrated from fit.

Original-Bug: 57348
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507119
Original-Revision: 101f75583079b733a8d3e06e9884406f2e87f08f
GitOrigin-RevId: e899478e8cc710b2fe68378f213e3c6092e6ea15
Change-Id: Icf60171df23bf0656b2e6986a463c660de7a7c0c