[roll] Roll fuchsia [ffx] Backend symbolization of proactive logs.
In this implementation we'll pass any log with a message that looks like
a context message to the `symbolizer` binary. We pass a known string
on a new line symbolizer to detect when a given input maps to zero or to
multiple lines of output.

Original-Fixed: 71770
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498351
Original-Revision: 1d0024d46ac2c6b942a2387f1d48a30903d08ea0
GitOrigin-RevId: c0c8c8bd87eb46c4eead89fa6f201a4d91c6f96e
Change-Id: I22cfa0cc4dbf894b13acb6735d7c1569c4411923