[roll] Roll fuchsia [fit] Allow lib/fit in the kernel
Update the README text to account for the fit-promise split.
Also remove an obsolete public_configs setting that is not needed.

This includes a single migration of an fbl::MakeAutoCall use in the
kernel to fit::defer as demonstration fit in the kernel works now.

Original-Fixed: 57350
Original-Bug: 57355
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507082
Original-Revision: b89444c0e8a15ca828895d92b982d655fe7e8b25
GitOrigin-RevId: 73d70bdc47723e0cc3635e97dc83fda9513ccad8
Change-Id: I1212a37002fed372748c689e8a1eeade236eb94c