[roll] Roll fuchsia [triage] Make select support multiple sources in one.
Migrations of triage selectors to a new location is currently difficult.
It requires creating a parallel set of rules that check both locations a
metric may appear. This CL extends "select" clauses to support a vector
of locations, the first of which is used for the actual value.

Original-Fixed: 62002
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440114
Original-Revision: f32fcdd7f8daf69857e9409e28bf1a50d1d6690c
GitOrigin-RevId: 55617d516e3781ab3b22f659f57e759f27eb5e7d
Change-Id: Ibcc267f567fe6921cac9c30f0f9dca29c4a8a7b5