[roll] Roll fuchsia [blobfs] Make fsck continue after errors
Fsck used to early exit when it detected corruption in the backup
superblock, which is not very useful since it hides the other issues
with the image. Change this to let fsck proceed through all checks
before returning whether the image is valid or not.

Tested: unit.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506798
Original-Revision: c454c5545ef80ed2e9fd0c8905fdbf03c8eddf51
GitOrigin-RevId: 340234a1eef993ec375fcd5b9f8e5b3dfc27b153
Change-Id: Id687b8d259690999d7044d5d16c9d5b0667b6f5d