[roll] Roll fuchsia [fidlgen][cpp] Improved name handling
Collapse type and decl handling into a single set of types.

Rename `DeclVariant` to `Name` and `DeclName` to `NameVariants`. It
turns out that in practice it's most common to name types and
declarations that are unique to a particular bindings variant, so the
API now centers on that.

Names themselves are represented in a tree of `namePart`s rather than an
array of strings. It's often very useful to be able to access a class'
name for itself - what it would call its constructor or destructor - and
due to how template specializations are named we need to introduce a
little more structure.

The main test for this change is that it doesn't cause a change in the
goldens.

Original-Bug: 72798

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506922
Original-Revision: b5ff2c4ccc005f24936db3d1df270058d406938c
GitOrigin-RevId: 6f68381cc83f12d198cc9e21a03373a1a5dd6aad
Change-Id: I8c01e11ae5af17ab5d9893b077373c5d7c78ba2b