[roll] Roll fuchsia [bt][gap] Restore bonds before starting scan
When the bonds are not restored before we start a passive scan (via
applying the configuration), restoring bonds can fail if we detect a
peer we are bonded to before the bonds are restored.

Restore the bonds before applying the config, ensuring this can't
happen.

Original-Bug: b/182420131
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507200
Original-Revision: 816dba4ebb7453a90c5f268cfedbcbca7406c368
GitOrigin-RevId: 40f44618107bf719647e58f260a694886d531273
Change-Id: I8dd7b82b4e2663a4628b724958ba8d91910986a0