[roll] Roll fuchsia [volume_image]: Blobfs Adapter.
Provides an adapter that generates the necessary metadata for processing
a blobfs image into a FvmDescriptor.

Added integration test for the image creation from input images.
Currently only blobfs.

This uncovered an OOB bug in the writing image code, added a regression
test for that as well.

Also uncovered a bug when writing a fvm image, where the mapping.source
offset was never applied when reading. This was not caught by the
existing tests, because the fake data never changed source to a non zero
value. No regression test added, but updated tests to trigger the bug,
and applied the fix.

Test: storage-volume-image-tests,
storage-volume-image-adapter-integration-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505560
Original-Revision: 47f84f17e40cdf59c29b17b7cac27958331f8513
GitOrigin-RevId: 39d25e618444a2895cc58962e860f223f693488f
Change-Id: I6794469b140d4046d5aca568024a752c54e07036