[roll] Roll fuchsia [zbitl] Remove individual Header trait implementations
To ease the introduction of bootfs support, we aim to make StorageTraits
ZBI-agnostic, which in turn means removing the Header() trait. As things
are spelled today, it is almost the case that we can write a unified
Header() method in zbitl::View, leveraging the most optimized Read()
method to read out a zbi_header_t; there are two issues:
(1) For the one-shot Read signature, it is generally UB to reinterpret
the returned std::byte* as a zbi_header_t*;
(2) Though the mapping VMO storage type admits one-shot reading, we
don't want to use it when reading headers, as we do not presume that
mapping costs are amortized by sequential access patterns when doing
item iteration (only when accessing payloads).

In this change, we write a unified Header() method and address these two
issues as follows:
(1): we update the storage trait spec so that storage is expected to be
`__STDCPP_DEFAULT_NEW_ALIGNMENT__`-aligned and that one-shot reading is
parametrized so that it can return return cpp20::span<T> (instead of
just `cpp20::span<const std::byte>`) if `T` is similarly aligned:
one can deduce from this that an item offset is then always
`alignof(T)`-aligned (no UB) (plus the storage backend now does all the
work);
(2): we further update the one-shot Read signature in adding a boolean
template parameter `LowLocality` giving whether we expect subsequent
reads of the adjacent data. If true, the storage backend is free to
implement the method differently or not implement it at all (as we do in
the case of the mapping VMO).
Putting these together, we implement the Header() method to directly
request of the backend to return a zbi_header_t* only after statically
determining whether it can one-shot read those with low locality.

Original-Bug: 68585
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506312
Original-Revision: c501689bc702bec88e59b05e991964a2aa41413d
GitOrigin-RevId: 4923deb4a965a442a5a9db8fc7a6d74f0ae8cc7d
Change-Id: Iedecd7a0443fa0b2b71f354c4886b4440b33c4cb