[roll] Roll fuchsia [ffx] ensure version information is generated
Previously the script did not fail when the version information could
not be gathered, and a recent change to the integration repository
resulted in HEAD being ambiguous to git show in some user scenarios.
We now rev-parse and then show that commit to generate our string,
whcih is more in-line wiht the behavior that the Zircon git hash
collection uses.

The dependencies are changed to only depend on HEAD as well, as a
small speculative change to try to somewhat reduce the frequency of
needless incremental rebuilds in infrastructure. This again shares
behavior with the Zircon git hash computation. A more complete
solution should follow as described in the associated bug.

Original-Bug: 73152
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507325
Original-Revision: 9d2cd2ae542cf809c0d4d88686e5691fe6606e94
GitOrigin-RevId: cbcc6370a3ee6b0c18d19c5d3ab010b16f1e7332
Change-Id: Ibabec3e583ae4d9893ce7b2ff997a413f381f966