[roll] Roll fuchsia [component_manager] CapabilityProvider doesn't Task::detach()
CapabilityProvider is a trait that component manager modules can use to
implement builtin or framework capabilities. It is common for trait
implementations to spawn tasks with detach().

Instead of calling detach(), have CapabilityProvider return the task and
stash it in an appropriate scope. For scoped framework capabilities,
store it in the instance of that scope. This will cause the task to get
cancelled if the scope instance is destroyed. For builtin capabilities,
store the task in component manager's instance.

This is one step toward removing use of Task::detach() in component_manager.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501960
Original-Revision: 0c577e80b0e7222bc5f520ea0ada0e1b0939e6d7
GitOrigin-RevId: a26d216b8d28ae27cc58131364c2a70e830fa652
Change-Id: I75ee33c495d3784fbc2224904247542d1420ad3f