[roll] Roll fuchsia [block] Fix use-after-free in MessageGroup
MessageGroup was unsafely using `delete this` while holding a lock on
this, which resulted in a UAF when the lock guard went out of scope.

Fixed this by refactoring so we don't need to use `delete this` to clean
up the MessageGroup.

Original-Bug: 73044 # Fixed
Tested: core-unittest.cmx with ASAN enabled.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507199
Original-Revision: 2d04c6b0cb91d7602b133ffc7b2bf215d0a45eb7
GitOrigin-RevId: 4d3e7057635cb7a8241ee633b1e58d43c84732cd
Change-Id: I0e7266774e3404fa473070d0f2540042a0cdf855