[roll] Roll fuchsia [fx-set] Generate compile_commands.json + rust-project.json
The legacy version of `fx set` passes
`--export-compile-commands=default` and `--export-rust-project` to GN,
which I forgot to add to the new version of fx (I was mainly concerned
about checking that the new version produced the same GN args, and
neglected to check that it produced the same GN flags).

fint already has support for generating compile_commands.json via the
`generate_compdb` static.proto field, but we need to be able to specify
the "default" target to avoid the latency of generating compile commands
for the entire build graph, so I added a new `compdb_targets` field to
override the set of targets that were include in the compdb.

Original-Bug: 73231
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508038
Original-Revision: e0f55730f28504347260786b328e0fbc02aeb158
GitOrigin-RevId: 904a9e4d756b9b59a27915c2bb6fd363c529a3a2
Change-Id: I8d4c41a3a63f018b901ef5f17edbcc3488c02825