[roll] Roll fuchsia [kernel][vm] Clear split bits of forked markers
When performing a fast merge of a parent and remaining child page lists
we only need to update the pages that the removed VMO forked.
Previously markers in removed VMO would be skipped, when in reality a
marker existing indicates a page fork happened, same as if a page is
present. Failing to update the split bits in this case would result in
a consistency error, manifesting only if another clone is created in
the future, as demonstrated by the included test case.

Original-Bug: 72141
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507900
Original-Revision: 037300ca550d4af9b5cba97455da8a709a306654
GitOrigin-RevId: 986005f973c8ada4764295a6af0a8065a4104212
Change-Id: Id36ed4cd7011881b4afa7be95a13518f9516e362