[roll] Roll fuchsia [kernel][arm64][feature] save the ASID bit size
Instead of panicing when booting on an 8 bit asid cpu, save the feature
and print a warning instead.

It's still not supported to really run for any length of time on an 8
bit ASID cpu but it should at least make an attempt at booting.

Save the feature bits for future code to switch behavior based on it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507335
Original-Revision: 8eac9106e82fdcb72b09ecb8ff46f421886feebc
GitOrigin-RevId: 5ec77ef0b6a2b7f01c72103b33295897f5350856
Change-Id: Ie99505c09ea19d5390cb6dbc24ce045e926a6f45