[roll] Roll fuchsia [blobfs] Remove thread from paging test.
The paging test ran many blobfs commands on a background thread, while
simultaneously accessing Blobfs and Blobfs on the main thread. Blobfs is
not designed for this. Fortunately, the background thread seems to be
unneeded and we can run the loop on the main thread.

Original-Bug: 72754
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508065
Original-Revision: 12a9f451f7fb21de1b7fe34412c6d5cc9fbdec0b
GitOrigin-RevId: 3b78b5f540d93de22da54fe1fd26d0075687421b
Change-Id: Ie943bdd09596de45b901cb4789944e2e163513ff