[roll] Roll fuchsia [fx set] Rebuild fx-set after HEAD changes
If we only rebuild when JIRI_HEAD changes, `fx set` won't be rebuilt
after checking out an older version of fuchsia.git using raw `git
checkout`. This may result in a failure if the built version of `fx set`
is incompatible with the checked-out fuchsia.git revision.

To handle that case, rebuild whenever HEAD changes, rather than
only when JIRI_HEAD changes.

Note that using raw git to check out and build old versions of
fuchsia.git is not an officially supported workflow, as old versions of
fuchsia.git may not be compatible with up-to-date versions of other
repos and CIPD packages pinned in global integration. However, that
workflow is still useful and generally "just works" as long as you don't
go back in history more than a few days. So we might as well make it
less likely to break, as long as it's easy. It would also be confusing
for a tool built from source in fuchsia.git to not auto-update when
switching to a different version of its sources.

Two potential future improvements:
1. Skip rebuilding if HEAD is *ahead* of "origin/master" (or
   "origin/main" in the future), unless the --dev flag is set. This will
   avoid rebuilding in the case where the user has made local changes
   unrelated to "fx set".
2. Maintain a cache of old versions of `fx-set` keyed by revision to
   avoid rebuilding when switching back and forth between two branches
   multiple times. But then we'd have to deal with garbage collection
   and such, which I'd rather not think about unless it's absolutely
   necessary.

Original-Bug: 68465
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508219
Original-Revision: b4f4c0e13d595040505d6da50c4a6ab8467aef74
GitOrigin-RevId: eba79a1d3ab16f2b6ddcc2cc76459d6168742ee2
Change-Id: I065dcfc9887079e6df8687bfc15de5a765ba28ce