[roll] Roll external/github.com/catapult-project/catapult devil: Restart the adb server after adb commands timeout
This CL makes changes so that it restarts the adb server after
an adb command times out, which was suggested as a work around in
crbug.com/1056214. Try builders running with this change show
that retries of the adb command are able to run successfully.

Test: Created test catapult roll in crrev.com/c/2787579. Try builders
run for the roll shows that restarting the adb server after a timeout
allowed subsequent re-runs of the command to run successfully.

Original-Bug:chromium:1056214
Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2787707
Original-Revision: c8248182dbaaf337c1f61c77bb54b8e00be4d2e0
GitOrigin-RevId: 9a251c8b406e698f977590de737e642eca730a0e
Change-Id: Ie8feb11f3ec528088171df1210808b0cb8d8339e