[roll] Roll infra/recipes [fint] Move more zbi_test logic into `fint build`
This logic handles figuring out which images to build based on the
contents of zbi_tests.json. It's translated from the zbi_test.py recipe:
https://fuchsia.googlesource.com/infra/recipes/+/HEAD/recipes/zbi_test.py#290

It's not load-bearing until I start switching zbi_test builders to use
`fint build`. This also means I can't test it on any optional zbi_test
builders in CQ, but I ran it locally with a fint params file for a
couple zbi_test builders and confirmed that it selects the same Ninja
targets as the recipe code.

This logic is pretty ugly because it needs to work around some
shortcomings of the build system. IMO we should move some of this work
into GN in the long term, but moving it into fint first will make it
easier to eventually migrate it into the build system without requiring
painful soft transitions with recipes.

Original-Bug: 67861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507284
GitOrigin-RevId: 3339b4a626ab747aaaac68977e14679d7e6aabaa
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/508216
Original-Revision: d6fb062df0468845fad8465bce91771185e84c4a
Change-Id: I699a8904df54e9c1fcc343fb16bca494a7cccd4c