[roll] Roll fuchsia [llcpp] Implementation of fidl_encode_iovec that linearizes to a buffer
This CL implements a somewhat trivial version of fidl_encode_ivoec
that does not use any optimizations to avoid copying and instead
always linearizes the input to a buffer and a single iovec entry.

The idea is to first implement the iovec interface and migrate to
it and then later add optimizations. In the next step, LLCPP objects
will switch over to use fidl_encode_iovec.

Original-Bug: 66977
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501980
Original-Revision: 0110adb3fefce0a295670c89f3a7322185bd061f
GitOrigin-RevId: 0a3fc65bb7258729d250d8e2b24a5d14f37b81b4
Change-Id: Id560649463fbc9fa76c223ade036e2df7173acb2