[roll] Roll fuchsia [blobfs] Fix some problems with new pager.
Blobfs::Create will actually make the Blobs. But before this patch, the
Vfs was not set on the Blobfs object until after Create was complete.
This meant that none of the created blobs got a Vfs pointer, resulting
in the pager failing.

Creation is changed to take the Vfs as an input pointer. A set function
is added on the Runner object to link the two sides (this was done in
the reverse order before).

Additionally, PagedVfs creation is improved to avoid reentrancy created
by the paged blobs being deleted after the destructor was complete.

Original-Bug: 51111

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506563
Original-Revision: d3c209de315cfe126e72d9e323ee20a9d3315051
GitOrigin-RevId: 412d5d7381566354c3c6aa3390ac8af33b27e66b
Change-Id: If429dae1aa168fc9ae652ba3c0a0ea6de0ca9264