[roll] Roll fuchsia [kernel][stacks] Allocate a vmo only large enough to hold the stack
The inner kstack allocation and mapping routine was previously always
allocating DEFAULT_STACK_SIZE vmos instead of the size passed in on the
particular stack type descriptor. In the case of shadow call stacks on
arm this means the vmo was actually sized at 8192 bytes instead of 4096
as was needed.

This was basically harmless because only the mapped portion of the vmo
is faulted in as the mapping is created. However some future code may
want to fiddle with additional types and would immediately notice if the
size was greater than DEFAULT_STACK_SIZE.

Cleaned up the kstack implementation to encapsulate the inner stacks
into a repeated structure.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418117
Original-Revision: ee7b5e550831563fe80f8ab201a598ab3dc61069
GitOrigin-RevId: 3f44ee6d9c2449bba7881d27f89d0d6fecedb706
Change-Id: I34fdbb2c62bb13592b5e33f7a74c5e1d6c1887af