[roll] Roll fuchsia [testrunner] Collect data sinks from component v2 tests.
This adds a RemoteDir field to the DataSinkReference to distinguish
between the two different directories where the data will be written to
for v1 versus v2 tests. It also collects all v2 data sinks at the end of
all tests because sometimes it takes some time for the summary.json to
get written to so we may fail to get the data sinks for the first test
if we try to read the summary.json before it's written.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/503538
Original-Revision: c8fc4a14cbf8f0c3094510193643a7faebbeab4d
GitOrigin-RevId: 18778081ed054cbe079d9b5695a0cc3f3517a82b
Change-Id: Ic771b8f2984dc2a4445f1af3a75abe381e87fd19