[roll] Roll fuchsia [bt][hfp] Identify calls by index instead of phone number.
In the HFP procedures, Calls are referred to by an assigned call index
instead of by the phone number of the remote party.

Introduces a new data structure, `CallList` which has special operations
for insertion, retrieval, and deletion so that the index associated with
a given call conforms to the standard.

Original-Bug: 64563
Test: Add unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497388
Original-Revision: 8dd0f4a242a50aa9d70babca39ce642782ad0135
GitOrigin-RevId: cdd62b5e7304e159b4f566b7966518986a4fe341
Change-Id: I6c1c3711bd2c9eb9d06f02ea88eb0e978fc9fed8