[roll] Roll fuchsia [kernel][pmm] Atomic for free count
The free_count_ is currently read racily without a lock. Although this
is unlikely to cause a correctness issue, it is still undefined
behavior. In this change it is made an atomic with all atomic accesses
using the relaxed memory ordering, with the locking itself left alone.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508438
Original-Revision: 16343879b01414cffe265a345b0737456493904c
GitOrigin-RevId: 73ed4f054b254f022621c6501d72426a738ae506
Change-Id: Ic0d8b1b8b00adc619c4d75fdeb2b865cf7da343d