[roll] Roll fuchsia [fidlgen][cpp] #ifdef __Fuchsia__ handling
It turns out that the namespace stack was only useful and used for
managing when entering and leaving sections of generated code that may
only be compiled on Fuchsia. This has now been reworked so that the
PushNamespace and PopNamespace template functions become IfdefFuchsia
and EndifFuchsia template functions and handle generating ifdef and
endif preprocessor directives (if needed).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508238
Original-Revision: 84efd4409f4f5ac97c56835aefae4247ba87cb22
GitOrigin-RevId: c5818b5c6585a092af8d5a17e8471df793ede171
Change-Id: Id1de50521df1c911dba0c157daa21d66326e7c59