[roll] Roll fuchsia [bt][hfp] Add format type method to Number.
The format of the number should be associated with the Number itself.
Number is moved to a submodule so that Numbers must be constructed
with methods which can verify the format and modify the string value.

All Numbers must include enclosing double-quotes which are added.

NUMBER_FORMAT is chosen to be within the range 128-143 which indicates
an international or national number that should be presented as is.

Test: Added unit tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508549
Original-Revision: 5c8b61c7099daf8dfea3599a61c45f24d3c484b3
GitOrigin-RevId: 8b6d0541db84795dedda198cd02f100bce4b714d
Change-Id: I214c7da02ef1e00a059694a8ce0af12f76d313b7